@@ -1,4 +1,3 @@
-# pylint: disable=too-many-lines
 import time
 from collections.abc import Sequence
 from functools import partial
@@ -13,14 +12,12 @@ from qiskit.primitives import (
     BaseSamplerV2,
 )
 from qiskit.quantum_info import SparsePauliOp
-from qiskit_algorithms.optimizers import COBYLA, L_BFGS_B, SLSQP, SPSA
 
 from slowquant.molecularintegrals.integralfunctions import (
     one_electron_integral_transform,
     two_electron_integral_transform,
 )
 from slowquant.qiskit_interface.interface import QuantumInterface
-from slowquant.qiskit_interface.optimizers import RotoSolve
 from slowquant.unitary_coupled_cluster.density_matrix import (
     ReducedDenstiyMatrix,
     get_electronic_energy,
@@ -28,15 +25,15 @@ from slowquant.unitary_coupled_cluster.density_matrix import (
 )
 from slowquant.unitary_coupled_cluster.fermionic_operator import FermionicOperator
 from slowquant.unitary_coupled_cluster.operators import Epq, hamiltonian_0i_0a
+from slowquant.unitary_coupled_cluster.optimizers import Optimizers
 
 
-class WaveFunction:
+class WaveFunctionCircuit:
     def __init__(
         self,
-        num_spin_orbs: int,
         num_elec: int,
         cas: Sequence[int],
-        c_orthonormal: np.ndarray,
+        mo_coeffs: np.ndarray,
         h_ao: np.ndarray,
         g_ao: np.ndarray,
         quantum_interface: QuantumInterface,
@@ -45,11 +42,10 @@ class WaveFunction:
         """Initialize for UCC wave function.
 
         Args:
-            num_spin_orbs: Number of spin orbitals.
             num_elec: Number of electrons.
             cas: CAS(num_active_elec, num_active_orbs),
                  orbitals are counted in spatial basis.
-            c_orthonormal: Initial orbital coefficients.
+            mo_coeffs: Initial orbital coefficients.
             h_ao: One-electron integrals in AO for Hamiltonian.
             g_ao: Two-electron integrals in AO.
             quantum_interface: QuantumInterface.
@@ -58,10 +54,12 @@ class WaveFunction:
         if len(cas) != 2:
             raise ValueError(f"cas must have two elements, got {len(cas)} elements.")
         if isinstance(quantum_interface.ansatz, QuantumCircuit):
-            print("WARNING: A QI with a custom Ansatz was passed. VQE will only work with COBYLA optimizer.")
-        self._c_orthonormal = c_orthonormal
-        self.h_ao = h_ao
-        self.g_ao = g_ao
+            print(
+                "WARNING: A QI with a custom Ansatz was passed. VQE will only work with COBYLA and COBYQA optimizer."
+            )
+        self._c_mo = mo_coeffs
+        self._h_ao = h_ao
+        self._g_ao = g_ao
         self.inactive_spin_idx = []
         self.virtual_spin_idx = []
         self.active_spin_idx = []
@@ -71,9 +69,8 @@ class WaveFunction:
         self.active_occ_spin_idx_shifted = []
         self.active_unocc_spin_idx_shifted = []
         self.num_elec = num_elec
-        self.num_spin_orbs = num_spin_orbs
-        self.num_orbs = num_spin_orbs // 2
-        self._include_active_kappa = include_active_kappa
+        self.num_spin_orbs = 2 * len(h_ao)
+        self.num_orbs = len(h_ao)
         self.num_active_elec = 0
         self.num_active_spin_orbs = 0
         self.num_inactive_spin_orbs = 0
@@ -84,6 +81,7 @@ class WaveFunction:
         self._rdm4 = None
         self._h_mo = None
         self._g_mo = None
+        self._energy_elec: float | None = None
         self.do_trace_corrected = True
         active_space = []
         orbital_counter = 0
@@ -101,7 +99,7 @@ class WaveFunction:
             else:
                 self.inactive_spin_idx.append(i)
                 self.num_inactive_spin_orbs += 1
-        for i in range(num_elec, num_spin_orbs):
+        for i in range(num_elec, self.num_spin_orbs):
             if i in active_space:
                 self.active_spin_idx.append(i)
                 self.active_unocc_spin_idx.append(i)
@@ -120,7 +118,7 @@ class WaveFunction:
         self.num_inactive_orbs = self.num_inactive_spin_orbs // 2
         self.num_active_orbs = self.num_active_spin_orbs // 2
         self.num_virtual_orbs = self.num_virtual_spin_orbs // 2
-        # Contruct spatial idx
+        # Construct spatial idx
         self.inactive_idx: list[int] = []
         self.virtual_idx: list[int] = []
         self.active_idx: list[int] = []
@@ -142,61 +140,41 @@ class WaveFunction:
             if idx // 2 not in self.active_unocc_idx:
                 self.active_unocc_idx.append(idx // 2)
         # Find non-redundant kappas
-        self.kappa = []
+        self._kappa = []
         self.kappa_idx = []
         self.kappa_no_activeactive_idx = []
         self.kappa_no_activeactive_idx_dagger = []
-        self.kappa_redundant = []
         self.kappa_redundant_idx = []
         self._kappa_old = []
-        self._kappa_redundant_old = []
         # kappa can be optimized in spatial basis
         for p in range(0, self.num_orbs):
             for q in range(p + 1, self.num_orbs):
                 if p in self.inactive_idx and q in self.inactive_idx:
-                    self.kappa_redundant.append(0.0)
-                    self._kappa_redundant_old.append(0.0)
-                    self.kappa_redundant_idx.append([p, q])
+                    self.kappa_redundant_idx.append((p, q))
                     continue
                 if p in self.virtual_idx and q in self.virtual_idx:
-                    self.kappa_redundant.append(0.0)
-                    self._kappa_redundant_old.append(0.0)
-                    self.kappa_redundant_idx.append([p, q])
+                    self.kappa_redundant_idx.append((p, q))
                     continue
                 if not include_active_kappa:
                     if p in self.active_idx and q in self.active_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
-                        continue
-                if include_active_kappa:
-                    if p in self.active_occ_idx and q in self.active_occ_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
-                        continue
-                    if p in self.active_unocc_idx and q in self.active_unocc_idx:
-                        self.kappa_redundant.append(0.0)
-                        self._kappa_redundant_old.append(0.0)
-                        self.kappa_redundant_idx.append([p, q])
+                        self.kappa_redundant_idx.append((p, q))
                         continue
                 if not (p in self.active_idx and q in self.active_idx):
-                    self.kappa_no_activeactive_idx.append([p, q])
-                    self.kappa_no_activeactive_idx_dagger.append([q, p])
-                self.kappa.append(0.0)
+                    self.kappa_no_activeactive_idx.append((p, q))
+                    self.kappa_no_activeactive_idx_dagger.append((q, p))
+                self._kappa.append(0.0)
                 self._kappa_old.append(0.0)
-                self.kappa_idx.append([p, q])
-        # HF like orbital rotation indecies
+                self.kappa_idx.append((p, q))
+        # HF like orbital rotation indices
         self.kappa_hf_like_idx = []
         for p in range(0, self.num_orbs):
             for q in range(p + 1, self.num_orbs):
                 if p in self.inactive_idx and q in self.virtual_idx:
-                    self.kappa_hf_like_idx.append([p, q])
+                    self.kappa_hf_like_idx.append((p, q))
                 elif p in self.inactive_idx and q in self.active_unocc_idx:
-                    self.kappa_hf_like_idx.append([p, q])
+                    self.kappa_hf_like_idx.append((p, q))
                 elif p in self.active_occ_idx and q in self.virtual_idx:
-                    self.kappa_hf_like_idx.append([p, q])
-        self._energy_elec: float | None = None
+                    self.kappa_hf_like_idx.append((p, q))
         # Setup Qiskit stuff
         self.QI = quantum_interface
         self.QI.construct_circuit(
@@ -204,45 +182,42 @@ class WaveFunction:
         )
 
     @property
-    def c_orthonormal(self) -> np.ndarray:
-        """Get orthonormalization coefficients (MO coefficients).
+    def kappa(self) -> list[float]:
+        """Get orbital rotation parameters."""
+        return self._kappa.copy()
 
-        Returns:
-            Orthonormalization coefficients.
-        """
-        return self._c_orthonormal
-
-    @c_orthonormal.setter
-    def c_orthonormal(self, c: np.ndarray) -> None:
-        """Set orthonormalization coefficients.
+    @kappa.setter
+    def kappa(self, k: list[float]) -> None:
+        """Set orbital rotation parameters, and move current expansion point.
 
         Args:
-            c: Orthonormalization coefficients.
+            k: orbital rotation parameters.
         """
         self._h_mo = None
         self._g_mo = None
         self._energy_elec = None
-        self._c_orthonormal = c
+        self._kappa = k.copy()
+        # Move current expansion point.
+        self._c_mo = self.c_mo
+        self._kappa_old = self.kappa
 
     @property
-    def c_trans(self) -> np.ndarray:
-        """Get orbital coefficients.
+    def c_mo(self) -> np.ndarray:
+        """Get molecular orbital coefficients.
 
         Returns:
-            Orbital coefficients.
+            Molecular orbital coefficients.
         """
-        kappa_mat = np.zeros_like(self._c_orthonormal)
+        kappa_mat = np.zeros_like(self._c_mo)
         if len(self.kappa) != 0:
-            if np.max(np.abs(self.kappa)) > 0.0:
-                for kappa_val, (p, q) in zip(self.kappa, self.kappa_idx):
-                    kappa_mat[p, q] = kappa_val
-                    kappa_mat[q, p] = -kappa_val
-        if len(self.kappa_redundant) != 0:
-            if np.max(np.abs(self.kappa_redundant)) > 0.0:
-                for kappa_val, (p, q) in zip(self.kappa_redundant, self.kappa_redundant_idx):
-                    kappa_mat[p, q] = kappa_val
-                    kappa_mat[q, p] = -kappa_val
-        return np.matmul(self._c_orthonormal, scipy.linalg.expm(-kappa_mat))
+            # The MO transformation is calculated as a difference between current kappa and kappa old.
+            # This is to make the moving of the expansion point to work with SciPy optimization algorithms.
+            # Resetting kappa to zero would mess with any algorithm that has any memory f.x. BFGS.
+            if np.max(np.abs(np.array(self.kappa) - np.array(self._kappa_old))) > 0.0:
+                for kappa_val, kappa_old, (p, q) in zip(self.kappa, self._kappa_old, self.kappa_idx):
+                    kappa_mat[p, q] = kappa_val - kappa_old
+                    kappa_mat[q, p] = -(kappa_val - kappa_old)
+        return np.matmul(self._c_mo, scipy.linalg.expm(-kappa_mat))
 
     @property
     def h_mo(self) -> np.ndarray:
@@ -252,7 +227,7 @@ class WaveFunction:
             One-electron Hamiltonian integrals in MO basis.
         """
         if self._h_mo is None:
-            self._h_mo = one_electron_integral_transform(self.c_trans, self.h_ao)
+            self._h_mo = one_electron_integral_transform(self.c_mo, self._h_ao)
         return self._h_mo
 
     @property
@@ -263,11 +238,11 @@ class WaveFunction:
             Two-electron Hamiltonian integrals in MO basis.
         """
         if self._g_mo is None:
-            self._g_mo = two_electron_integral_transform(self.c_trans, self.g_ao)
+            self._g_mo = two_electron_integral_transform(self.c_mo, self._g_ao)
         return self._g_mo
 
     @property
-    def ansatz_parameters(self) -> list[float]:
+    def thetas(self) -> list[float]:
         """Getter for ansatz parameters.
 
         Returns:
@@ -275,8 +250,8 @@ class WaveFunction:
         """
         return self.QI.parameters
 
-    @ansatz_parameters.setter
-    def ansatz_parameters(self, parameters: list[float]) -> None:
+    @thetas.setter
+    def thetas(self, parameters: list[float]) -> None:
         """Setter for ansatz paramters.
 
         Args:
@@ -343,7 +318,7 @@ class WaveFunction:
 
     @property
     def rdm1(self) -> np.ndarray:
-        r"""Calcuate one-electron reduced density matrix.
+        r"""Calculate one-electron reduced density matrix.
 
         The trace condition is enforced:
 
@@ -375,7 +350,7 @@ class WaveFunction:
 
     @property
     def rdm2(self) -> np.ndarray:
-        r"""Calcuate two-electron reduced density matrix.
+        r"""Calculate two-electron reduced density matrix.
 
         The trace condition is enforced:
 
@@ -434,7 +409,7 @@ class WaveFunction:
 
     @property
     def rdm3(self) -> np.ndarray:
-        r"""Calcuate three-electron reduced density matrix.
+        r"""Calculate three-electron reduced density matrix.
 
         The trace condition is enforced:
 
@@ -507,7 +482,7 @@ class WaveFunction:
 
     @property
     def rdm4(self) -> np.ndarray:
-        r"""Calcuate four-electron reduced density matrix.
+        r"""Calculate four-electron reduced density matrix.
 
         The trace condition is enforced:
 
@@ -748,7 +723,7 @@ class WaveFunction:
         return self._rdm4
 
     def precalc_rdm_paulis(self, rdm_order: int) -> None:
-        """Pre-calculate all Paulis used to contruct RDMs up to a certain order.
+        """Pre-calculate all Paulis used to construct RDMs up to a certain order.
 
         This utilizes the saving feature in QuantumInterface when using the Sampler primitive.
         If saving is turned up in QuantumInterface this function will do nothing but waste device time.
@@ -765,7 +740,7 @@ class WaveFunction:
         if rdm_order > 4:
             raise ValueError(f"Precalculation only supported up to order 4 got {rdm_order}")
         if rdm_order < 1:
-            raise ValueError(f"Precalculation need atleast an order of 1 got {rdm_order}")
+            raise ValueError(f"Precalculation need at least an order of 1 got {rdm_order}")
         cumulated_paulis = None
         if rdm_order >= 1:
             self._rdm1 = None
@@ -846,7 +821,7 @@ class WaveFunction:
         Args:
             overlap_integral: Overlap integral in AO basis.
         """
-        S_ortho = one_electron_integral_transform(self.c_trans, overlap_integral)
+        S_ortho = one_electron_integral_transform(self.c_mo, overlap_integral)
         one = np.identity(len(S_ortho))
         diff = np.abs(S_ortho - one)
         print("Max ortho-normal diff:", np.max(diff))
@@ -859,9 +834,7 @@ class WaveFunction:
             Electronic energy.
         """
         if self._energy_elec is None:
-            H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
-            H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
-            self._energy_elec = self.QI.quantum_expectation_value(H)
+            self._energy_elec = self._calc_energy_elec()
         return self._energy_elec
 
     def _calc_energy_elec(self) -> float:
@@ -873,169 +846,94 @@ class WaveFunction:
         H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
         H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
         energy_elec = self.QI.quantum_expectation_value(H)
-
         return energy_elec
 
-    def _get_hamiltonian(self) -> FermionicOperator:
-        """Return electronic Hamiltonian as FermionicOperator.
-
-        Returns:
-            FermionicOperator.
-        """
-        H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
-        H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
-
-        return H
-
-    def run_vqe_2step(
+    def run_wf_optimization_2step(
         self,
-        ansatz_optimizer: str,
+        optimizer_name: str,
         orbital_optimization: bool = False,
-        tol: float = 1e-8,
+        tol: float = 1e-10,
         maxiter: int = 1000,
         is_silent_subiterations: bool = False,
     ) -> None:
-        """Run VQE of wave function."""
-        global iteration  # pylint: disable=global-variable-undefined
-        global start  # pylint: disable=global-variable-undefined
+        """Run two step optimization of wave function.
 
-        if isinstance(self.QI.ansatz, QuantumCircuit) and not ansatz_optimizer.lower() == "cobyla":
-            raise ValueError("Custom Ansatz in QI only works with COBYLA as optimizer")
-
-        def print_progress(x, energy_func, silent: bool) -> None:
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                x: Wave function parameters.
-                energy_func: Function to calculate energy.
-                silent: Supress print.
-            """
-            global iteration  # pylint: disable=global-variable-undefined
-            global start  # pylint: disable=global-variable-undefined
-            time_str = f"{time.time() - start:7.2f}"  # type: ignore [name-defined] # pylint: disable=used-before-assignment
-            if not silent:
-                e_str = f"{energy_func(x):3.16f}"
-                print(
-                    f"--------{str(iteration + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}"  # type: ignore [name-defined] # pylint: disable=used-before-assignment
-                )
-            iteration += 1  # type: ignore
-            start = time.time()  # type: ignore
-
-        def print_progress_SPSA(
-            ___,
-            theta,  # pylint: disable=unused-argument
-            f_val,
-            _,
-            __,
-            silent: bool,
-        ) -> None:
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                theta: Wave function parameters.
-                f_val: Function value at theta.
-                silent: Supress print.
-            """
-            global iteration  # pylint: disable=global-variable-undefined
-            global start  # pylint: disable=global-variable-undefined
-            time_str = f"{time.time() - start:7.2f}"  # type: ignore
-            e_str = f"{f_val:3.12f}"
-            if not silent:
-                print(f"--------{str(iteration + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}")  # type: ignore
-            iteration += 1  # type: ignore
-            start = time.time()  # type: ignore
-
+        Args:
+            optimizer_name: Name of optimizer.
+            orbital_optimization: Perform orbital optimization.
+            tol: Convergence tolerance.
+            maxiter: Maximum number of iterations.
+            is_silent_subiterations: Silence subiterations.
+        """
+        if isinstance(self.QI.ansatz, QuantumCircuit) and not optimizer_name.lower() in ("cobyla", "cobyqa"):
+            raise ValueError("Custom Ansatz in QI only works with COBYLA and COBYQA as optimizer.")
+        print("### Parameters information:")
+        if orbital_optimization:
+            print(f"### Number kappa: {len(self.kappa)}")
+        print(f"### Number theta: {len(self.thetas)}")
         e_old = 1e12
         print("Full optimization")
         print("Iteration # | Iteration time [s] | Electronic energy [Hartree]")
         for full_iter in range(0, int(maxiter)):
             full_start = time.time()
-            iteration = 0  # type: ignore
-            start = time.time()  # type: ignore
 
             # Do ansatz optimization
             if not is_silent_subiterations:
                 print("--------Ansatz optimization")
                 print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
-            H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
-            H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
             energy_theta = partial(
-                calc_energy_theta,
-                operator=H,
-                quantum_interface=self.QI,
+                self._calc_energy_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
             )
-            gradient_theta = partial(ansatz_parameters_gradient, operator=H, quantum_interface=self.QI)
-            if ansatz_optimizer.lower() == "slsqp":
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_theta, silent=is_silent_subiterations
-                )
-                optimizer = SLSQP(maxiter=maxiter, ftol=tol, callback=print_progress_)
-            elif ansatz_optimizer.lower() == "slsqp_nograd":
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_theta, silent=is_silent_subiterations
-                )
-                optimizer = SLSQP(maxiter=maxiter, ftol=tol, callback=print_progress_)
-            elif ansatz_optimizer.lower() == "l_bfgs_b":
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_theta, silent=is_silent_subiterations
-                )
-                optimizer = L_BFGS_B(maxiter=maxiter, tol=tol, callback=print_progress_)
-            elif ansatz_optimizer.lower() == "cobyla":
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_theta, silent=is_silent_subiterations
-                )
-                optimizer = COBYLA(maxiter=maxiter, tol=tol, callback=print_progress_)
-            elif ansatz_optimizer.lower() == "rotosolve":
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_theta, silent=is_silent_subiterations
-                )
-                optimizer = RotoSolve(
-                    self.QI.grad_param_R,
-                    self.QI.param_names,
-                    maxiter=maxiter,
-                    tol=tol,
-                    callback=print_progress_,
-                )
-            elif ansatz_optimizer.lower() == "spsa":
-                print("WARNING: Convergence tolerence cannot be set for SPSA; using qiskit default")
-                print_progress_SPSA_ = partial(print_progress_SPSA, silent=is_silent_subiterations)
-                optimizer = SPSA(maxiter=maxiter, callback=print_progress_SPSA_)
-            else:
-                raise ValueError(f"Unknown optimizer: {ansatz_optimizer}")
-            if ansatz_optimizer.lower() == "slsqp_nograd":
-                res = optimizer.minimize(energy_theta, self.ansatz_parameters)
-            else:
-                res = optimizer.minimize(energy_theta, self.ansatz_parameters, jac=gradient_theta)
-            self.ansatz_parameters = res.x.tolist()
+            gradient_theta = partial(
+                self._calc_gradient_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+            )
+            optimizer = Optimizers(
+                energy_theta,
+                optimizer_name,
+                grad=gradient_theta,
+                maxiter=maxiter,
+                tol=tol,
+                is_silent=is_silent_subiterations,
+            )
+            res = optimizer.minimize(
+                self.thetas,
+                extra_options={"R": self.QI.grad_param_R, "param_names": self.QI.param_names},
+            )
+            self.thetas = res.x.tolist()
 
             if orbital_optimization and len(self.kappa) != 0:
-                iteration = 0  # type: ignore
-                start = time.time()  # type: ignore
                 if not is_silent_subiterations:
                     print("--------Orbital optimization")
                     print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
                 energy_oo = partial(
-                    calc_energy_oo,
-                    wf=self,
+                    self._calc_energy_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
                 )
-                gradiet_oo = partial(
-                    orbital_rotation_gradient,
-                    wf=self,
+                gradient_oo = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
                 )
 
-                print_progress_ = partial(
-                    print_progress, energy_func=energy_oo, silent=is_silent_subiterations
+                optimizer = Optimizers(
+                    energy_oo,
+                    "l-bfgs-b",
+                    grad=gradient_oo,
+                    maxiter=maxiter,
+                    tol=tol,
+                    is_silent=is_silent_subiterations,
                 )
-                optimizer = L_BFGS_B(maxiter=maxiter, tol=tol, callback=print_progress_)
-                res = optimizer.minimize(energy_oo, [0.0] * len(self.kappa_idx), jac=gradiet_oo)
+                res = optimizer.minimize([0.0] * len(self.kappa_idx))
                 for i in range(len(self.kappa)):  # pylint: disable=consider-using-enumerate
-                    self.kappa[i] = 0.0
+                    self._kappa[i] = 0.0
                     self._kappa_old[i] = 0.0
-                for i in range(len(self.kappa_redundant)):  # pylint: disable=consider-using-enumerate
-                    self.kappa_redundant[i] = 0.0
-                    self._kappa_redundant_old[i] = 0.0
             else:
-                # If theres is no orbital optimization, then the algorithm is already converged.
+                # If there is no orbital optimization, then the algorithm is already converged.
                 e_new = res.fun
                 if orbital_optimization and len(self.kappa) == 0:
                     print(
@@ -1052,269 +950,184 @@ class WaveFunction:
             e_old = e_new
         self._energy_elec = e_new
 
-    def run_vqe_1step(
+    def run_wf_optimization_1step(
         self,
         optimizer_name: str,
         orbital_optimization: bool = False,
-        tol: float = 1e-8,
+        tol: float = 1e-10,
         maxiter: int = 1000,
     ) -> None:
-        """Run VQE of wave function."""
-        if not orbital_optimization:
-            raise ValueError("Does only work with orbital optimization right now")
-        global iteration  # pylint: disable=global-variable-undefined
-        global start  # pylint: disable=global-variable-undefined
-        iteration = 0  # type: ignore
-        start = time.time()  # type: ignore
+        """Run one step optimization of wave function.
 
-        if isinstance(self.QI.ansatz, QuantumCircuit) and not optimizer_name.lower() == "cobyla":
-            raise ValueError("Custom Ansatz in QI only works with COBYLA as optimizer")
-
-        def print_progress(x, energy_func) -> None:
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                x: Wave function parameters.
-                energy_func: Function to calculate energy.
-            """
-            global iteration  # pylint: disable=global-variable-undefined
-            global start  # pylint: disable=global-variable-undefined
-            time_str = f"{time.time() - start:7.2f}"  # type: ignore
-            e_str = f"{energy_func(x):3.12f}"
-            print(f"{str(iteration + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}")  # type: ignore
-            iteration += 1  # type: ignore
-            start = time.time()  # type: ignore
-
-        def print_progress_SPSA(
-            ___,
-            theta,  # pylint: disable=unused-argument
-            f_val,
-            _,
-            __,
-        ) -> None:
-            """Print progress during energy minimization of wave function.
-
-            Args:
-                theta: Wave function parameters.
-                f_val: Function value at theta.
-            """
-            global iteration  # pylint: disable=global-variable-undefined
-            global start  # pylint: disable=global-variable-undefined
-            time_str = f"{time.time() - start:7.2f}"  # type: ignore
-            e_str = f"{f_val:3.12f}"
-            print(f"{str(iteration + 1).center(11)} | {time_str.center(18)} | {e_str.center(27)}")  # type: ignore
-            iteration += 1  # type: ignore
-            start = time.time()  # type: ignore
-
-        print("Iteration # | Iteration time [s] | Electronic energy [Hartree]")
-        energy_both = partial(
-            calc_energy_both,
-            wf=self,
-        )
-        gradient_both = partial(
-            calc_gradient_both,
-            wf=self,
-        )
-        if optimizer_name.lower() == "slsqp":
-            print_progress_ = partial(print_progress, energy_func=energy_both)
-            optimizer = SLSQP(maxiter=maxiter, ftol=tol, callback=print_progress_)
-        elif optimizer_name.lower() == "l_bfgs_b":
-            print_progress_ = partial(print_progress, energy_func=energy_both)
-            optimizer = L_BFGS_B(maxiter=maxiter, tol=tol, callback=print_progress_)
-        elif optimizer_name.lower() == "cobyla":
-            print_progress_ = partial(print_progress, energy_func=energy_both)
-            optimizer = COBYLA(maxiter=maxiter, tol=tol, callback=print_progress_)
-        elif optimizer_name.lower() == "rotosolve":
+        Args:
+            optimizer_name: Name of optimizer.
+            orbital_optimization: Perform orbital optimization.
+            tol: Convergence tolerance.
+            maxiter: Maximum number of iterations.
+        """
+        if isinstance(self.QI.ansatz, QuantumCircuit) and not optimizer_name.lower() in ("cobyla", "cobyqa"):
+            raise ValueError("Custom Ansatz in QI only works with COBYLA and COBYQA as optimizer.")
+        print("### Parameters information:")
+        if orbital_optimization:
+            print(f"### Number kappa: {len(self.kappa)}")
+        print(f"### Number theta: {len(self.thetas)}")
+        if optimizer_name.lower() == "rotosolve":
             if orbital_optimization and len(self.kappa) != 0:
                 raise ValueError(
-                    "Cannot use rotosolve together with orbital optimization in the one-step solver."
+                    "Cannot use RotoSolve together with orbital optimization in the one-step solver."
                 )
-            print_progress_ = partial(print_progress, energy_func=energy_both)
-            optimizer = RotoSolve(
-                self.QI.grad_param_R, self.QI.param_names, maxiter=maxiter, tol=tol, callback=print_progress_
-            )
-        elif optimizer_name.lower() == "spsa":
-            print("WARNING: Convergence tolerence cannot be set for SPSA; using qiskit default")
-            optimizer = SPSA(maxiter=maxiter, callback=print_progress_SPSA)
+
+        print("--------Iteration # | Iteration time [s] | Electronic energy [Hartree]")
+        if orbital_optimization:
+            if len(self.thetas) > 0:
+                energy = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=True,
+                )
+                gradient = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=True,
+                    kappa_optimization=True,
+                )
+            else:
+                energy = partial(
+                    self._calc_energy_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
+                gradient = partial(
+                    self._calc_gradient_optimization,
+                    theta_optimization=False,
+                    kappa_optimization=True,
+                )
         else:
-            raise ValueError(f"Unknown optimizer: {optimizer_name}")
-        parameters = self.kappa + self.ansatz_parameters
-        res = optimizer.minimize(energy_both, parameters, jac=gradient_both)
-        self.ansatz_parameters = res.x[len(self.kappa) :].tolist()
-        for i in range(len(self.kappa)):  # pylint: disable=consider-using-enumerate
-            self.kappa[i] = 0.0
-            self._kappa_old[i] = 0.0
-        for i in range(len(self.kappa_redundant)):  # pylint: disable=consider-using-enumerate
-            self.kappa_redundant[i] = 0.0
-            self._kappa_redundant_old[i] = 0.0
+            energy = partial(
+                self._calc_energy_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+            )
+            gradient = partial(
+                self._calc_gradient_optimization,
+                theta_optimization=True,
+                kappa_optimization=False,
+            )
+        if orbital_optimization:
+            if len(self.thetas) > 0:
+                parameters = self.kappa + self.thetas
+            else:
+                parameters = self.kappa
+        else:
+            parameters = self.thetas
+        optimizer = Optimizers(energy, optimizer_name, grad=gradient, maxiter=maxiter, tol=tol)
+        res = optimizer.minimize(
+            parameters, extra_options={"R": self.QI.grad_param_R, "param_names": self.QI.param_names}
+        )
+        if orbital_optimization:
+            self.thetas = res.x[len(self.kappa) :].tolist()
+            for i in range(len(self.kappa)):  # pylint: disable=consider-using-enumerate
+                self._kappa[i] = 0.0
+                self._kappa_old[i] = 0.0
+        else:
+            self.thetas = res.x.tolist()
         self._energy_elec = res.fun
 
+    def _calc_energy_optimization(
+        self, parameters: list[float], theta_optimization: bool, kappa_optimization: bool
+    ) -> float:
+        """Calculate electronic energy.
 
-def calc_energy_theta(
-    parameters: list[float], operator: FermionicOperator, quantum_interface: QuantumInterface
-) -> float:
-    """Calculate electronic energy using expectation values.
+        Args:
+            parameters: Ansatz and orbital rotation parameters.
+            theta_optimization: Doing theta optimization.
+            kappa_optimization: Doing kappa optimization.
 
-    Args:
-        paramters: Ansatz paramters.
-        operator: Hamiltonian operator.
-        quantum_interface: QuantumInterface.
+        Returns:
+            Electronic energy.
+        """
+        num_kappa = 0
+        if kappa_optimization:
+            num_kappa = len(self.kappa_idx)
+            self.kappa = parameters[:num_kappa]
+        if theta_optimization:
+            self.thetas = parameters[num_kappa:]
+            # Build operator
+            H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
+            H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
+            return self.QI.quantum_expectation_value(H)
+        # RDM is more expensive than evaluation of the Hamiltonian.
+        # Thus only construct these if orbital-optimization is turned on,
+        # since the RDMs will be reused in the oo gradient calculation.
+        rdms = ReducedDenstiyMatrix(
+            self.num_inactive_orbs,
+            self.num_active_orbs,
+            self.num_virtual_orbs,
+            rdm1=self.rdm1,
+            rdm2=self.rdm2,
+        )
+        return get_electronic_energy(rdms, self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
 
-    Returns:
-        Electronic energy.
-    """
-    quantum_interface.parameters = parameters
-    return quantum_interface.quantum_expectation_value(operator)
+    def _calc_gradient_optimization(
+        self, parameters: list[float], theta_optimization: bool, kappa_optimization: bool
+    ) -> np.ndarray:
+        """Calculate electronic gradient.
 
+        Args:
+            parameters: Ansatz and orbital rotation parameters.
+            theta_optimization: Doing theta optimization.
+            kappa_optimization: Doing kappa optimization.
 
-def calc_energy_oo(kappa: list[float], wf: WaveFunction) -> float:
-    """Calculate electronic energy using RDMs.
-
-    Args:
-        kappa: Orbital rotation parameters.
-        wf: Wave function object.
-
-    Returns:
-        Electronic energy.
-    """
-    kappa_mat = np.zeros_like(wf.c_orthonormal)
-    for kappa_val, (p, q) in zip(
-        np.array(kappa) - np.array(wf._kappa_old), wf.kappa_idx  # pylint: disable=protected-access
-    ):
-        kappa_mat[p, q] = kappa_val
-        kappa_mat[q, p] = -kappa_val
-    if len(wf.kappa_redundant) != 0:
-        if np.max(np.abs(wf.kappa_redundant)) > 0.0:
-            for kappa_val, (p, q) in zip(
-                np.array(wf.kappa_redundant)
-                - np.array(wf._kappa_redundant_old),  # pylint: disable=protected-access
-                wf.kappa_redundant_idx,
-            ):
-                kappa_mat[p, q] = kappa_val
-                kappa_mat[q, p] = -kappa_val
-    c_trans = np.matmul(wf.c_orthonormal, scipy.linalg.expm(-kappa_mat))
-    wf._kappa_old = kappa.copy()  # pylint: disable=protected-access
-    wf._kappa_redundant_old = wf.kappa_redundant.copy()  # pylint: disable=protected-access
-    # Moving expansion point of kappa
-    wf.c_orthonormal = c_trans
-    rdms = ReducedDenstiyMatrix(
-        wf.num_inactive_orbs,
-        wf.num_active_orbs,
-        wf.num_active_orbs,
-        rdm1=wf.rdm1,
-        rdm2=wf.rdm2,
-    )
-    energy = get_electronic_energy(rdms, wf.h_mo, wf.g_mo, wf.num_inactive_orbs, wf.num_active_orbs)
-    return energy
-
-
-def calc_energy_both(parameters, wf) -> float:
-    """Calculate electronic energy.
-
-    Args:
-        parameters: Ansatz and orbital rotation parameters.
-        wf: Wave function object.
-
-    Returns:
-        Electronic energy.
-    """
-    kappa = parameters[: len(wf.kappa)]
-    theta = parameters[len(wf.kappa) :]
-    assert len(theta) == len(wf.ansatz_parameters)
-    # Do orbital partial
-    kappa_mat = np.zeros_like(wf.c_orthonormal)
-    for kappa_val, (p, q) in zip(
-        np.array(kappa) - np.array(wf._kappa_old), wf.kappa_idx  # pylint: disable=protected-access
-    ):
-        kappa_mat[p, q] = kappa_val
-        kappa_mat[q, p] = -kappa_val
-    if len(wf.kappa_redundant) != 0:
-        if np.max(np.abs(wf.kappa_redundant)) > 0.0:
-            for kappa_val, (p, q) in zip(
-                np.array(wf.kappa_redundant)
-                - np.array(wf._kappa_redundant_old),  # pylint: disable=protected-access
-                wf.kappa_redundant_idx,
-            ):
-                kappa_mat[p, q] = kappa_val
-                kappa_mat[q, p] = -kappa_val
-    c_trans = np.matmul(wf.c_orthonormal, scipy.linalg.expm(-kappa_mat))
-    wf._kappa_old = kappa.copy()  # pylint: disable=protected-access
-    wf._kappa_redundant_old = wf.kappa_redundant.copy()  # pylint: disable=protected-access
-    # Moving expansion point of kappa
-    wf.c_orthonormal = c_trans
-    # Build operator
-    wf.ansatz_parameters = theta.copy()  # Reset rdms
-    H = hamiltonian_0i_0a(wf.h_mo, wf.g_mo, wf.num_inactive_orbs, wf.num_active_orbs)
-    H = H.get_folded_operator(wf.num_inactive_orbs, wf.num_active_orbs, wf.num_virtual_orbs)
-    return wf.QI.quantum_expectation_value(H)
+        Returns:
+            Electronic gradient.
+        """
+        num_kappa = 0
+        gradient = np.zeros(len(parameters))
+        num_kappa = 0
+        if kappa_optimization:
+            num_kappa = len(self.kappa_idx)
+            self.kappa = parameters[:num_kappa]
+        if theta_optimization:
+            self.thetas = parameters[num_kappa:]
+        if kappa_optimization:
+            rdms = ReducedDenstiyMatrix(
+                self.num_inactive_orbs,
+                self.num_active_orbs,
+                self.num_virtual_orbs,
+                rdm1=self.rdm1,
+                rdm2=self.rdm2,
+            )
+            gradient[:num_kappa] = get_orbital_gradient(
+                rdms, self.h_mo, self.g_mo, self.kappa_idx, self.num_inactive_orbs, self.num_active_orbs
+            )
+        if theta_optimization:
+            H = hamiltonian_0i_0a(self.h_mo, self.g_mo, self.num_inactive_orbs, self.num_active_orbs)
+            H = H.get_folded_operator(self.num_inactive_orbs, self.num_active_orbs, self.num_virtual_orbs)
+            for i in range(len(parameters[num_kappa:])):  # pylint: disable=consider-using-enumerate
+                R = self.QI.grad_param_R[self.QI.param_names[i]]
+                e_vals_grad = _get_energy_evals_for_grad(H, self.QI, parameters, i, R)
+                grad = 0.0
+                for j, mu in enumerate(list(range(1, 2 * R + 1))):
+                    x_mu = (2 * mu - 1) / (2 * R) * np.pi
+                    grad += e_vals_grad[j] * (-1) ** (mu - 1) / (4 * R * (np.sin(1 / 2 * x_mu)) ** 2)
+                gradient[num_kappa + i] = grad
+        return gradient
 
 
-def orbital_rotation_gradient(
-    placeholder,  # pylint: disable=unused-argument
-    wf,
-) -> np.ndarray:
-    """Calcuate electronic gradient with respect to orbital rotations.
-
-    Args:
-        placeholder: Placeholder for kappa parameters, these are fetched OOP style instead.
-        wf: Wave function object.
-
-    Return:
-        Electronic gradient with respect to orbital rotations.
-    """
-    rdms = ReducedDenstiyMatrix(
-        wf.num_inactive_orbs,
-        wf.num_active_orbs,
-        wf.num_active_orbs,
-        rdm1=wf.rdm1,
-        rdm2=wf.rdm2,
-    )
-    gradient = get_orbital_gradient(
-        rdms, wf.h_mo, wf.g_mo, wf.kappa_idx, wf.num_inactive_orbs, wf.num_active_orbs
-    )
-    return gradient
-
-
-def ansatz_parameters_gradient(
-    parameters: list[float], operator: FermionicOperator, quantum_interface: QuantumInterface
-) -> np.ndarray:
-    r"""Calculate gradient with respect to ansatz parameters.
-
-    Args:
-        parameters: Ansatz parameters.
-        operator: Operator which the derivative is with respect to.
-        quantum_interface: Interface to call quantum device.
-
-    Returns:
-        Gradient with repsect to ansatz parameters.
-    """
-    gradient = np.zeros(len(parameters))
-    for i in range(len(parameters)):  # pylint: disable=consider-using-enumerate
-        R = quantum_interface.grad_param_R[quantum_interface.param_names[i]]
-        e_vals_grad = get_energy_evals_for_grad(operator, quantum_interface, parameters, i, R)
-        grad = 0.0
-        for j, mu in enumerate(list(range(1, 2 * R + 1))):
-            x_mu = (2 * mu - 1) / (2 * R) * np.pi
-            grad += e_vals_grad[j] * (-1) ** (mu - 1) / (4 * R * (np.sin(1 / 2 * x_mu)) ** 2)
-        gradient[i] = grad
-    return gradient
-
-
-def get_energy_evals_for_grad(
+def _get_energy_evals_for_grad(
     operator: FermionicOperator,
     quantum_interface: QuantumInterface,
     parameters: list[float],
     idx: int,
     R: int,
 ) -> list[float]:
-    r"""Get energy evaluations needed for the gradient calculation.
+    """Get energy evaluations needed for the gradient calculation.
 
-    The gradient formula is defined for x=0,
-    so x_shift is used to shift ensure we can get the energy in the point we actually want.
+    The gradient formula is defined for x=0.
+    The x_shift variable is used to shift the energy function, such that current parameter value is in zero.
 
     Args:
         operator: Operator which the derivative is with respect to.
-        parameters: Paramters.
+        parameters: Parameters.
         idx: Parameter idx.
         R: Parameter to control we get the needed points.
 
@@ -1329,25 +1142,3 @@ def get_energy_evals_for_grad(
         x[idx] = x_mu + x_shift
         e_vals.append(quantum_interface.quantum_expectation_value(operator, custom_parameters=x))
     return e_vals
-
-
-def calc_gradient_both(parameters: list[float], wf: WaveFunction) -> np.ndarray:
-    """Calculate electronic gradient.
-
-    Args:
-        parameters: Ansatz and orbital rotation parameters.
-        wf: Wave function object.
-
-    Returns:
-        Electronic gradient.
-    """
-    gradient = np.zeros(len(parameters))
-    theta = parameters[len(wf.kappa) :]
-    assert len(theta) == len(wf.ansatz_parameters)
-    kappa_grad = orbital_rotation_gradient(0, wf)
-    gradient[: len(wf.kappa)] = kappa_grad
-    H = hamiltonian_0i_0a(wf.h_mo, wf.g_mo, wf.num_inactive_orbs, wf.num_active_orbs)
-    H = H.get_folded_operator(wf.num_inactive_orbs, wf.num_active_orbs, wf.num_virtual_orbs)
-    theta_grad = ansatz_parameters_gradient(theta, H, wf.QI)
-    gradient[len(wf.kappa) :] = theta_grad
-    return gradient
