@@ -8,14 +8,14 @@ from qiskit_nature.second_q.mappers.fermionic_mapper import FermionicMapper
 
 from slowquant.qiskit_interface.operators_circuits import (
     double_excitation,
+    sa_single_excitation,
     single_excitation,
-    single_sa_excitation,
-    tups_double,
-    tups_single,
 )
 from slowquant.unitary_coupled_cluster.util import (
+    iterate_pair_t2,
     iterate_pair_t2_generalized,
     iterate_t1,
+    iterate_t1_sa,
     iterate_t1_sa_generalized,
     iterate_t2,
 )
@@ -29,7 +29,7 @@ def tUPS(
 ) -> tuple[QuantumCircuit, dict[str, int]]:
     """Create tUPS ansatz.
 
-    #. 10.1103/PhysRevResearch.6.023300
+    #. 10.1103/PhysRevResearch.6.023300 (tUPS)
     #. 10.1088/1367-2630/ac2cb3 (QNP)
 
     Ansatz Options:
@@ -63,8 +63,8 @@ def tUPS(
     else:
         do_qnp = False
 
-    if not isinstance(mapper, JordanWignerMapper):
-        raise ValueError(f"tUPS only implemented for JW mapper, got: {type(mapper)}")
+    if not isinstance(mapper, JordanWignerMapper) and do_pp:
+        raise ValueError(f"pp-tUPS only implemented for JW mapper, got: {type(mapper)}")
     if do_pp and np.sum(num_elec) != num_orbs:
         raise ValueError(
             f"pp-tUPS only implemented for number of electrons and number of orbitals being the same, got: ({np.sum(num_elec)}, {num_orbs}), (elec, orbs)"
@@ -74,25 +74,21 @@ def tUPS(
     else:
         skip_last_singles = False
 
-    num_qubits = 2 * num_orbs  # qc.num_qubits
-    if do_pp:
-        qc = QuantumCircuit(num_qubits)
-        for p in range(0, 2 * num_orbs):
-            if p % 2 == 0:
-                qc.x(p)
-    else:
-        qc = HartreeFock(num_orbs, num_elec, mapper)
+    qc = HartreeFock(num_orbs, (0, 0), mapper)  # empty circuit with qubit number based on mapper
     grad_param_R = {}
     idx = 0
+    # Layer loop
     for n in range(n_layers):
-        for p in range(0, num_orbs - 1, 2):
+        for p in range(0, num_orbs - 1, 2):  # first column of brick-wall
             if not do_qnp:
                 # First single
-                qc = tups_single(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+                qc = sa_single_excitation(p, p + 1, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
                 grad_param_R[f"p{idx:09d}"] = 4
                 idx += 1
             # Double
-            qc = tups_double(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+            qc = double_excitation(
+                2 * p, 2 * p + 1, 2 * p + 2, 2 * p + 3, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper
+            )
             grad_param_R[f"p{idx:09d}"] = 2
             idx += 1
             # Second single
@@ -101,23 +97,25 @@ def tUPS(
                 # Here the layer is only one block, thus,
                 # the last single excitation is earlier than expected.
                 continue
-            qc = tups_single(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+            qc = sa_single_excitation(p, p + 1, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
             grad_param_R[f"p{idx:09d}"] = 4
             idx += 1
-        for p in range(1, num_orbs - 1, 2):
+        for p in range(1, num_orbs - 1, 2):  # second column of brick-wall
             if not do_qnp:
                 # First single
-                qc = tups_single(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+                qc = sa_single_excitation(p, p + 1, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
                 grad_param_R[f"p{idx:09d}"] = 4
                 idx += 1
             # Double
-            qc = tups_double(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+            qc = double_excitation(
+                2 * p, 2 * p + 1, 2 * p + 2, 2 * p + 3, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper
+            )
             grad_param_R[f"p{idx:09d}"] = 2
             idx += 1
             # Second single
             if n + 1 == n_layers and skip_last_singles:
                 continue
-            qc = tups_single(p, num_orbs, qc, Parameter(f"p{idx:09d}"))
+            qc = sa_single_excitation(p, p + 1, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
             grad_param_R[f"p{idx:09d}"] = 4
             idx += 1
     return qc, grad_param_R
@@ -129,30 +127,61 @@ def fUCC(
     mapper: FermionicMapper,
     ansatz_options: dict[str, Any],
 ) -> tuple[QuantumCircuit, dict[str, int]]:
-    """Create factorized UCCSD ansatz.
+    """Create factorized UCC ansatz.
 
-    #. 10.1103/PhysRevA.102.062612
+    #. 10.1103/PhysRevA.102.062612 (efficient circuits for JW)
+    #. 10.1021/acs.jctc.8b01004 (k-UpCCGSD)
 
     Ansatz Options:
         * n_layers [int]: Number of layers.
+        * S [bool]: Add single excitations.
+        * SAS [bool]: Add spin-adapted single excitations.
+        * SAGS [bool]: Add generalized spin-adapted single excitations.
+        * D [bool]: Add double excitations.
+        * pD [bool]: Add pair double excitations.
+        * GpD [bool]: Add generalized pair double excitations.
 
     Args:
         num_orbs: Number of spatial orbitals.
         num_elec: Number of alpha and beta electrons.
-        mapper: Fermioinc to qubit mapper.
+        mapper: Fermionic to qubit mapper.
         ansatz_options: Ansatz options.
 
     Returns:
-        Factorized UCCSD ansatz circuit and R parameters needed for gradients.
+        Factorized UCC ansatz circuit and R parameters needed for gradients.
     """
-    valid_options = ("n_layers",)
+    valid_options = ("n_layers", "S", "D", "SAGS", "pD", "GpD", "SAS")
     for option in ansatz_options:
         if option not in valid_options:
             raise ValueError(f"Got unknown option for fUCC, {option}. Valid options are: {valid_options}")
-    if not isinstance(mapper, JordanWignerMapper):
-        raise ValueError(f"fUCC only implemented for JW mapper, got: {type(mapper)}")
     if "n_layers" not in ansatz_options.keys():
         raise ValueError("fUCC require the option 'n_layers'")
+    do_S = False
+    do_SAS = False
+    do_SAGS = False
+    do_D = False
+    do_pD = False
+    do_GpD = False
+    if "S" in ansatz_options.keys():
+        if ansatz_options["S"]:
+            do_S = True
+    if "SAS" in ansatz_options.keys():
+        if ansatz_options["SAS"]:
+            do_SAS = True
+    if "SAGS" in ansatz_options.keys():
+        if ansatz_options["SAGS"]:
+            do_SAGS = True
+    if "D" in ansatz_options.keys():
+        if ansatz_options["D"]:
+            do_D = True
+    if "pD" in ansatz_options.keys():
+        if ansatz_options["pD"]:
+            do_pD = True
+    if "GpD" in ansatz_options.keys():
+        if ansatz_options["GpD"]:
+            do_GpD = True
+    if True not in (do_S, do_SAS, do_SAGS, do_D, do_pD, do_GpD):
+        raise ValueError("fUCC requires some excitations got none.")
     n_layers = ansatz_options["n_layers"]
     num_spin_orbs = 2 * num_orbs
     occ = []
@@ -164,75 +193,51 @@ def fUCC(
     for _ in range(num_spin_orbs - np.sum(num_elec)):
         unocc.append(idx)
         idx += 1
-    qc = HartreeFock(num_orbs, num_elec, mapper)
+    qc = HartreeFock(num_orbs, (0, 0), mapper)  # empty circuit with qubit number based on mapper
     grad_param_R = {}
     idx = 0
+    # Layer loop
     for _ in range(n_layers):
-        for a, i in iterate_t1(occ, unocc):
-            qc = single_excitation(a, i, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-        for a, i, b, j in iterate_t2(occ, unocc):
-            qc = double_excitation(a, b, i, j, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
+        if do_S:
+            for a, i in iterate_t1(occ, unocc):
+                qc = single_excitation(i, a, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+        if do_SAS:
+            for a, i, _ in iterate_t1_sa(occ, unocc):
+                qc = sa_single_excitation(i, a, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
+        if do_SAGS:
+            for a, i, _ in iterate_t1_sa_generalized(num_orbs):
+                qc = sa_single_excitation(i, a, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
+        if do_D:
+            for a, i, b, j in iterate_t2(occ, unocc):
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+        if do_pD:
+            for a, i, b, j in iterate_pair_t2(occ, unocc):
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+        if do_GpD:
+            for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
     return qc, grad_param_R
 
 
-def kSAfUpCCGSD(
-    num_orbs: int, num_elec: tuple[int, int], mapper: FermionicMapper, ansatz_options: dict[str, Any]
-) -> tuple[QuantumCircuit, dict[str, int]]:
-    """Create modified k-UpCCGSD ansatz.
-
-    The ansatz has been modified to use spin-adapted singlet single excitation operators.
-
-    #. 10.1021/acs.jctc.8b01004
-
-    Ansatz Options:
-        * n_layers [int]: Number of layers.
-
-    Args:
-        num_orbs: Number of spatial orbitals.
-        num_elec: Number of alpha and beta electrons.
-        mapper: Fermioinc to qubit mapper.
-        ansatz_options: Ansatz options.
-
-    Returns:
-        Modified k-UpCCGSD ansatz.
-    """
-    valid_options = ("n_layers",)
-    for option in ansatz_options:
-        if option not in valid_options:
-            raise ValueError(
-                f"Got unknown option for kSAfUpCCGSD, {option}. Valid options are: {valid_options}"
-            )
-    if "n_layers" not in ansatz_options.keys():
-        raise ValueError("kSAfUpCCGSD require the option 'n_layers'")
-    if not isinstance(mapper, JordanWignerMapper):
-        raise ValueError(f"kSAfUpCCGSD only implemented for JW mapper, got: {type(mapper)}")
-    n_layers = ansatz_options["n_layers"]
-    qc = HartreeFock(num_orbs, num_elec, mapper)
-    grad_param_R = {}
-    idx = 0
-    for _ in range(n_layers):
-        for a, i, _ in iterate_t1_sa_generalized(num_orbs):
-            qc = single_sa_excitation(a, i, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 4
-            idx += 1
-        for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
-            qc = double_excitation(a, b, i, j, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-    return qc, grad_param_R
-
-
-def dUCCSD(
+def SDSfUCC(
     num_orbs: int,
     num_elec: tuple[int, int],
     mapper: FermionicMapper,
     ansatz_options: dict[str, Any],
 ) -> tuple[QuantumCircuit, dict[str, int]]:
-    r"""Create disentangled UCCSD ansatz with SDS ordering.
+    r"""Create SDS ordered factorized UCC.
 
     The operator ordering of this implementation is,
 
@@ -241,10 +246,15 @@ def dUCCSD(
         \exp\left(\theta_{ijab}\left(\hat{T}_{ijab}-\hat{T}_{ijab}^\dagger\right)\right)
         \exp\left(\theta_{ia}\left(\hat{T}_{ia}-\hat{T}_{ia}^\dagger\right)\right)\left|\text{CSF}\right>
 
-    #. 10.1063/1.5133059, Eq. 25, Eq. 35
+    #. 10.1063/1.5133059, Eq. 25, Eq. 35 (SDS)
+    #. 10.1103/PhysRevA.102.062612 (efficient circuits for JW)
+    #. 10.1021/acs.jctc.8b01004 (k-UpCCGSD)
 
     Ansatz Options:
         * n_layers [int]: Number of layers.
+        * D [bool]: Add double excitations.
+        * pD [bool]: Add pair double excitations.
+        * GpD [bool]: Add generalized pair double excitations.
 
     Args:
         num_orbs: Number of spatial orbitals.
@@ -253,16 +263,28 @@ def dUCCSD(
         ansatz_options: Ansatz options.
 
     Returns:
-        Disentangled UCCSD ansatz circuit and R parameters needed for gradients.
+        SDS ordered fUCC ansatz circuit and R parameters needed for gradients.
     """
-    valid_options = ("n_layers",)
+    valid_options = ("n_layers", "D", "pD", "GpD")
     for option in ansatz_options:
         if option not in valid_options:
-            raise ValueError(f"Got unknown option for dUCCSD, {option}. Valid options are: {valid_options}")
-    if not isinstance(mapper, JordanWignerMapper):
-        raise ValueError(f"dUCCSD only implemented for JW mapper, got: {type(mapper)}")
+            raise ValueError(f"Got unknown option for SDSfUCC, {option}. Valid options are: {valid_options}")
     if "n_layers" not in ansatz_options.keys():
-        raise ValueError("dUCCSD require the option 'n_layers'")
+        raise ValueError("SDSfUCC require the option 'n_layers'")
+    do_D = False
+    do_pD = False
+    do_GpD = False
+    if "D" in ansatz_options.keys():
+        if ansatz_options["D"]:
+            do_D = True
+    if "pD" in ansatz_options.keys():
+        if ansatz_options["pD"]:
+            do_pD = True
+    if "GpD" in ansatz_options.keys():
+        if ansatz_options["GpD"]:
+            do_GpD = True
+    if True not in (do_D, do_pD, do_GpD):
+        raise ValueError("SDSfUCC requires some excitations got none.")
     n_layers = ansatz_options["n_layers"]
     num_spin_orbs = 2 * num_orbs
     occ = []
@@ -274,69 +296,49 @@ def dUCCSD(
     for _ in range(num_spin_orbs - np.sum(num_elec)):
         unocc.append(idx)
         idx += 1
-    qc = HartreeFock(num_orbs, num_elec, mapper)
+    qc = HartreeFock(num_orbs, (0, 0), mapper)  # empty circuit with qubit number based on mapper
     grad_param_R = {}
     idx = 0
+    # Layer loop
     for _ in range(n_layers):
-        for a, i, b, j in iterate_t2(occ, unocc):
-            qc = single_excitation(a, i, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-            qc = double_excitation(a, b, i, j, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-            qc = single_excitation(b, j, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-    return qc, grad_param_R
-
-
-def kSAdUpCCGSD(
-    num_orbs: int, num_elec: tuple[int, int], mapper: FermionicMapper, ansatz_options: dict[str, Any]
-) -> tuple[QuantumCircuit, dict[str, int]]:
-    """Create modified k-UpCCGSD ansatz.
-
-    The ansatz have been modifed to use spin-adapted singet single excitation operators.
-    This ansatz has also been modified to follow the operator ordering as dUCCSD.
-
-    #. 10.1021/acs.jctc.8b01004
-    #. 10.1063/1.5133059, Eq. 25, Eq. 35 (for ordering)
-
-    Ansatz Options:
-        * n_layers [int]: Number of layers.
-
-    Args:
-        num_orbs: Number of spatial orbitals.
-        num_elec: Number of alpha and beta electrons.
-        mapper: Fermioinc to qubit mapper.
-        ansatz_options: Ansatz options.
-
-    Returns:
-        Modified k-UpCCGSD ansatz.
-    """
-    valid_options = ("n_layers",)
-    for option in ansatz_options:
-        if option not in valid_options:
-            raise ValueError(
-                f"Got unknown option for kSAdUpCCGSD, {option}. Valid options are: {valid_options}"
-            )
-    if "n_layers" not in ansatz_options.keys():
-        raise ValueError("kSAfUpCCGSD require the option 'n_layers'")
-    if not isinstance(mapper, JordanWignerMapper):
-        raise ValueError(f"kSAdUpCCGSD only implemented for JW mapper, got: {type(mapper)}")
-    n_layers = ansatz_options["n_layers"]
-    qc = HartreeFock(num_orbs, num_elec, mapper)
-    grad_param_R = {}
-    idx = 0
-    for _ in range(n_layers):
-        for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
-            qc = single_sa_excitation(a // 2, i // 2, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 4
-            idx += 1
-            qc = double_excitation(a, b, i, j, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 2
-            idx += 1
-            qc = single_sa_excitation(a // 2, i // 2, num_orbs, qc, Parameter(f"p{idx:09d}"))
-            grad_param_R[f"p{idx:09d}"] = 4
-            idx += 1
+        # Kind of D excitation determines indices for complete SDS block
+        if do_D:
+            for a, i, b, j in iterate_t2(occ, unocc):
+                if i % 2 == a % 2:
+                    qc = single_excitation(i, a, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                else:
+                    qc = single_excitation(i, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+                if i % 2 == a % 2:
+                    qc = single_excitation(j, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                else:
+                    qc = single_excitation(j, a, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+        if do_pD:
+            for a, i, b, j in iterate_pair_t2(occ, unocc):
+                qc = sa_single_excitation(i // 2, a // 2, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+                qc = sa_single_excitation(i // 2, a // 2, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
+        if do_GpD:
+            for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
+                qc = sa_single_excitation(i // 2, a // 2, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
+                qc = double_excitation(i, j, a, b, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 2
+                idx += 1
+                qc = sa_single_excitation(i // 2, a // 2, num_orbs, qc, Parameter(f"p{idx:09d}"), mapper)
+                grad_param_R[f"p{idx:09d}"] = 4
+                idx += 1
     return qc, grad_param_R
