@@ -150,17 +150,37 @@ class quantumLR(quantumLRBaseClass):
             for j, qJ in enumerate(self.q_ops):
                 for i, GI in enumerate(self.G_ops):
                     # Make A
-                    val = self.wf.QI.quantum_expectation_value(
-                        (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs)
-                    ) - self.wf.QI.quantum_expectation_value(
-                        (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
+                    val = (
+                        self.wf.QI.quantum_expectation_value(
+                            (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs)
+                        )
                     )
                     self.A[i + idx_shift, j] = self.A[j, i + idx_shift] = val
                     # Make B
-                    val = self.wf.QI.quantum_expectation_value(
-                        (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs)
-                    ) - self.wf.QI.quantum_expectation_value(
-                        (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    val = (
+                        self.wf.QI.quantum_expectation_value(
+                            (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                        )
+                        - 1
+                        / 2
+                        * self.wf.QI.quantum_expectation_value(
+                            (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                        )
                     )
                     self.B[i + idx_shift, j] = self.B[j, i + idx_shift] = val
 
@@ -170,7 +190,9 @@ class quantumLR(quantumLRBaseClass):
                 # Make A
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = (
                     self.wf.QI.quantum_expectation_value(
-                        double_commutator(GI.dagger, self.H_0i_0a, GJ).get_folded_operator(*self.orbs)
+                        double_commutator(
+                            GI.dagger, self.H_0i_0a, GJ, do_symmetrized=True
+                        ).get_folded_operator(*self.orbs)
                     )
                 )
                 # Make B
@@ -247,6 +269,9 @@ class quantumLR(quantumLRBaseClass):
                     + self.wf.QI.op_to_qbit(
                         (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs)
                     ).paulis.to_labels()
+                    + self.wf.QI.op_to_qbit(
+                        (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs)
+                    ).paulis.to_labels()
                 )
                 A[i + idx_shift][j] = A[j][i + idx_shift] = val
                 # Make B
@@ -257,6 +282,9 @@ class quantumLR(quantumLRBaseClass):
                     + self.wf.QI.op_to_qbit(
                         (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
                     ).paulis.to_labels()
+                    + self.wf.QI.op_to_qbit(
+                        (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs)
+                    ).paulis.to_labels()
                 )
                 B[i + idx_shift][j] = B[j][i + idx_shift] = val
 
@@ -265,7 +293,9 @@ class quantumLR(quantumLRBaseClass):
             for i, GI in enumerate(self.G_ops[j:], j):
                 # Make A
                 A[i + idx_shift][j + idx_shift] = A[j + idx_shift][i + idx_shift] = self.wf.QI.op_to_qbit(
-                    double_commutator(GI.dagger, self.H_1i_1a, GJ).get_folded_operator(*self.orbs)
+                    double_commutator(GI.dagger, self.H_1i_1a, GJ, do_symmetrized=True).get_folded_operator(
+                        *self.orbs
+                    )
                 ).paulis.to_labels()
                 # Make B
                 B[i + idx_shift][j + idx_shift] = B[j + idx_shift][i + idx_shift] = self.wf.QI.op_to_qbit(
@@ -374,8 +404,15 @@ class quantumLR(quantumLRBaseClass):
                     self.wf.QI.quantum_variance(
                         (GI.dagger * self.H_1i_1a * qJ).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
                     )
-                    + self.wf.QI.quantum_variance(
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
                         (self.H_1i_1a * qJ * GI.dagger).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
+                    )
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
+                        (self.H_1i_1a * GI.dagger * qJ).get_folded_operator(*self.orbs), no_coeffs=no_coeffs
                     )
                 )
                 A[i + idx_shift, j] = A[j, i + idx_shift] = val
@@ -385,8 +422,16 @@ class quantumLR(quantumLRBaseClass):
                         (qJ.dagger * self.H_1i_1a * GI.dagger).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
-                    + self.wf.QI.quantum_variance(
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
                         (GI.dagger * qJ.dagger * self.H_1i_1a).get_folded_operator(*self.orbs),
+                        no_coeffs=no_coeffs,
+                    )
+                    + 1
+                    / 2
+                    * self.wf.QI.quantum_variance(
+                        (qJ.dagger * GI.dagger * self.H_1i_1a).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
                 )
@@ -398,7 +443,9 @@ class quantumLR(quantumLRBaseClass):
                 # Make A
                 A[i + idx_shift, j + idx_shift] = A[j + idx_shift, i + idx_shift] = np.sqrt(
                     self.wf.QI.quantum_variance(
-                        double_commutator(GI.dagger, self.H_0i_0a, GJ).get_folded_operator(*self.orbs),
+                        double_commutator(
+                            GI.dagger, self.H_0i_0a, GJ, do_symmetrized=True
+                        ).get_folded_operator(*self.orbs),
                         no_coeffs=no_coeffs,
                     )
                 )
@@ -441,9 +488,9 @@ class quantumLR(quantumLRBaseClass):
             rdm2=self.wf.rdm2,
         )
 
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(mux, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muy_op = one_elec_op_0i_0a(muy, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
         muz_op = one_elec_op_0i_0a(muz, self.wf.num_inactive_orbs, self.wf.num_active_orbs)
