@@ -16,7 +16,7 @@ from slowquant.unitary_coupled_cluster.fermionic_operator import FermionicOperat
 from slowquant.unitary_coupled_cluster.linear_response.lr_baseclass import (
     LinearResponseBaseClass,
 )
-from slowquant.unitary_coupled_cluster.operator_matrix import (
+from slowquant.unitary_coupled_cluster.operator_state_algebra import (
     expectation_value,
     propagate_state,
 )
@@ -135,12 +135,27 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     Hq_ket,
                     *self.index_info,
                 )
-                # - <0| H q Gd |0>
-                val -= expectation_value(
-                    qdH_ket,
-                    [],
-                    Gd_ket,
-                    *self.index_info,
+                # - 1/2<0| H q Gd |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        qdH_ket,
+                        [],
+                        Gd_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2<0| H Gd q |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [self.H_1i_1a * GI.dagger * qJ],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
                 )
                 self.A[i + idx_shift, j] = self.A[j, i + idx_shift] = val
                 # Make B
@@ -151,12 +166,27 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     Gd_ket,
                     *self.index_info,
                 )
-                # - <0| Gd qd H |0>
-                val -= expectation_value(
-                    G_ket,
-                    [],
-                    qdH_ket,
-                    *self.index_info,
+                # - 1/2*<0| Gd qd H |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        G_ket,
+                        [],
+                        qdH_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2*<0| qd Gd H |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [qJ.dagger * GI.dagger * self.H_1i_1a],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
                 )
                 self.B[i + idx_shift, j] = self.B[j, i + idx_shift] = val
         for j, GJ in enumerate(self.G_ops):
@@ -177,26 +207,56 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     HGJ_ket,
                     *self.index_info,
                 )
-                # - <0| GId GJ H |0>
-                val -= expectation_value(
-                    GI_ket,
-                    [],
-                    GJH_ket,
-                    *self.index_info,
-                )
-                # - <0| H GJ GId |0>
-                val -= expectation_value(
-                    GJdH_ket,
-                    [],
-                    GId_ket,
-                    *self.index_info,
-                )
                 # <0| GJ H GId |0>
                 val += expectation_value(
                     HGJd_ket,
                     [],
                     GId_ket,
                     *self.index_info,
+                )
+                # - 1/2<0| GId GJ H |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        GI_ket,
+                        [],
+                        GJH_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2*<0| H GJ GId |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        GJdH_ket,
+                        [],
+                        GId_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2*<0| GJ GId H |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        GJd_ket,
+                        [GI.dagger],
+                        H00_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2*<0| H GId GJ |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        H00_ket,
+                        [GI.dagger],
+                        GJ_ket,
+                        *self.index_info,
+                    )
                 )
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = val
                 # Make B
@@ -265,9 +325,9 @@ class LinearResponseUCC(LinearResponseBaseClass):
             self.wf.rdm1,
             rdm2=self.wf.rdm2,
         )
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(
             mux,
             self.wf.num_inactive_orbs,
