@@ -13,7 +13,7 @@ from slowquant.unitary_coupled_cluster.density_matrix import (
 from slowquant.unitary_coupled_cluster.linear_response.lr_baseclass import (
     LinearResponseBaseClass,
 )
-from slowquant.unitary_coupled_cluster.operator_matrix import (
+from slowquant.unitary_coupled_cluster.operator_state_algebra import (
     expectation_value,
     propagate_state,
 )
@@ -30,14 +30,12 @@ class LinearResponseUCC(LinearResponseBaseClass):
         self,
         wave_function: WaveFunctionUCC | WaveFunctionUPS,
         excitations: str,
-        do_approximate_hermitification: bool = False,
     ) -> None:
         """Initialize linear response by calculating the needed matrices.
 
         Args:
             wave_function: Wave function object.
             excitations: Which excitation orders to include in response.
-            do_approximate_hermitification: approximated method with BqG = 0 and AqG made Hermitian
         """
         super().__init__(wave_function, excitations)
 
@@ -48,10 +46,6 @@ class LinearResponseUCC(LinearResponseBaseClass):
             self.wf.num_active_orbs,
             self.wf.num_virtual_orbs,
         )
-
-        num_parameters = len(self.G_ops) + len(self.q_ops)
-        if do_approximate_hermitification:
-            self.B_tracked = np.zeros((num_parameters, num_parameters))
 
         rdms = ReducedDenstiyMatrix(
             self.wf.num_inactive_orbs,
@@ -126,47 +120,17 @@ class LinearResponseUCC(LinearResponseBaseClass):
                 )
         for j, qJ in enumerate(self.q_ops):
             UdHq_ket = propagate_state(["Ud", self.H_1i_1a * qJ], self.wf.ci_coeffs, *self.index_info)
-            UdqdH_ket = propagate_state(["Ud", qJ.dagger * self.H_1i_1a], self.wf.ci_coeffs, *self.index_info)
             for i, GI in enumerate(self.G_ops):
                 G_ket = propagate_state([GI], self.wf.csf_coeffs, *self.index_info)
-                if do_approximate_hermitification:
-                    # Make A
-                    # <CSF| Gd Ud H q |0>
-                    val = expectation_value(
-                        G_ket,
-                        [],
-                        UdHq_ket,
-                        *self.index_info,
-                    )
-                    # <CSF| Gd Ud qd H |0>
-                    tmp = expectation_value(
-                        G_ket,
-                        [],
-                        UdqdH_ket,
-                        *self.index_info,
-                    )
-                    self.A[j, i + idx_shift] = self.A[i + idx_shift, j] = val + tmp
-                    # Make B
-                    self.B_tracked[j, i + idx_shift] = self.B_tracked[i + idx_shift, j] = -tmp
-                else:
-                    # Make A
-                    # <CSF| Gd Ud H q |0>
-                    val = expectation_value(
-                        G_ket,
-                        [],
-                        UdHq_ket,
-                        *self.index_info,
-                    )
-                    self.A[j, i + idx_shift] = self.A[i + idx_shift, j] = val
-                    # Make B
-                    # - <CSF| Gd Ud qd H |0>
-                    val = -expectation_value(
-                        G_ket,
-                        [],
-                        UdqdH_ket,
-                        *self.index_info,
-                    )
-                    self.B[j, i + idx_shift] = self.B[i + idx_shift, j] = val
+                # Make A
+                # <CSF| Gd Ud H q |0>
+                val = expectation_value(
+                    G_ket,
+                    [],
+                    UdHq_ket,
+                    *self.index_info,
+                )
+                self.A[j, i + idx_shift] = self.A[i + idx_shift, j] = val
         for j, GJ in enumerate(self.G_ops):
             UdHUGJ_ket = propagate_state(
                 ["Ud", self.H_0i_0a, "U", GJ],
@@ -208,9 +172,9 @@ class LinearResponseUCC(LinearResponseBaseClass):
             self.wf.rdm1,
             rdm2=self.wf.rdm2,
         )
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(
             mux,
             self.wf.num_inactive_orbs,
