@@ -2,28 +2,6 @@ from collections.abc import Generator, Sequence
 from typing import Any
 
 import numpy as np
-import scipy.linalg
-
-
-def construct_integral_trans_mat(
-    c_orthonormal: np.ndarray, kappa: Sequence[float], kappa_idx: Sequence[Sequence[int]]
-) -> np.ndarray:
-    """Contruct orbital transformation matrix.
-
-    Args:
-        c_orthonormal: Initial orbital coefficients.
-        kappa: Orbital rotation parameters.
-        kappa_idx: Non-redundant orbital rotation parameters indices.
-
-    Returns:
-        Orbital transformation matrix.
-    """
-    kappa_mat = np.zeros_like(c_orthonormal)
-    for kappa_val, (p, q) in zip(kappa, kappa_idx):
-        kappa_mat[p, q] = kappa_val
-        kappa_mat[q, p] = -kappa_val
-    c_trans = np.matmul(c_orthonormal, scipy.linalg.expm(-kappa_mat))
-    return c_trans
 
 
 def iterate_t1_sa(
@@ -37,7 +15,7 @@ def iterate_t1_sa(
         active_unocc_idx: Indices of weakly occupied orbitals.
 
     Returns:
-        T1 operator iteration.
+        Spin-adapted T1 operator iteration.
     """
     for i in active_occ_idx:
         for a in active_unocc_idx:
@@ -56,7 +34,7 @@ def iterate_t2_sa(
         active_unocc_idx: Indices of weakly occupied orbitals.
 
     Returns:
-        T2 operator iteration.
+        Spin-adapted T2 operator iteration.
     """
     for idx_i, i in enumerate(active_occ_idx):
         for j in active_occ_idx[idx_i:]:
@@ -84,7 +62,7 @@ def iterate_t1_sa_generalized(
         num_orbs: Number of active spatial orbitals.
 
     Returns:
-        T1 operator iteration.
+        Generalized spin-adapted T1 operator iteration.
     """
     for i in range(num_orbs):
         for a in range(i + 1, num_orbs):
@@ -437,7 +415,7 @@ def iterate_pair_t2(
         active_unocc_idx: Indices of weakly occupied orbitals.
 
     Returns:
-        T2 operator iteration.
+        Pair T2 operator iteration.
     """
     for i in active_occ_idx:
         for a in active_unocc_idx:
@@ -447,13 +425,13 @@ def iterate_pair_t2(
 def iterate_pair_t2_generalized(
     num_orbs: int,
 ) -> Generator[tuple[int, int, int, int], None, None]:
-    """Iterate over pair T2 operators.
+    """Iterate over generalized pair T2 operators.
 
     Args:
         num_orbs: Number of active spatial orbitals.
 
     Returns:
-        T2 operator iteration.
+        Generlaized pair T2 operator iteration.
     """
     for i in range(num_orbs):
         for a in range(i + 1, num_orbs):
@@ -461,9 +439,11 @@ def iterate_pair_t2_generalized(
 
 
 class UccStructure:
+    __slots__ = ("excitation_indices", "excitation_operator_type", "n_params")
+
     def __init__(self) -> None:
         """Intialize the unitary coupled cluster ansatz structure."""
-        self.excitation_indicies: list[tuple[int, ...]] = []
+        self.excitation_indices: list[tuple[int, ...]] = []
         self.excitation_operator_type: list[str] = []
         self.n_params = 0
 
@@ -475,7 +455,7 @@ class UccStructure:
             active_unocc_idx: Active weakly occupied spatial orbital indices.
         """
         for a, i, _ in iterate_t1_sa(active_occ_idx, active_unocc_idx):
-            self.excitation_indicies.append((i, a))
+            self.excitation_indices.append((i, a))
             self.excitation_operator_type.append("sa_single")
             self.n_params += 1
 
@@ -487,7 +467,7 @@ class UccStructure:
             active_unocc_idx: Active weakly occupied spatial orbital indices.
         """
         for a, i, b, j, _, op_type in iterate_t2_sa(active_occ_idx, active_unocc_idx):
-            self.excitation_indicies.append((i, j, a, b))
+            self.excitation_indices.append((i, j, a, b))
             if op_type == 1:
                 self.excitation_operator_type.append("sa_double_1")
             elif op_type == 2:
@@ -502,7 +482,7 @@ class UccStructure:
             active_unocc_spin_idx: Active weakly occupied spin orbital indices.
         """
         for a, i, b, j, c, k in iterate_t3(active_occ_spin_idx, active_unocc_spin_idx):
-            self.excitation_indicies.append((i, j, k, a, b, c))
+            self.excitation_indices.append((i, j, k, a, b, c))
             self.excitation_operator_type.append("triple")
             self.n_params += 1
 
@@ -516,7 +496,7 @@ class UccStructure:
             active_unocc_spin_idx: Active weakly occupied spin orbital indices.
         """
         for a, i, b, j, c, k, d, l in iterate_t4(active_occ_spin_idx, active_unocc_spin_idx):
-            self.excitation_indicies.append((i, j, k, l, a, b, c, d))
+            self.excitation_indices.append((i, j, k, l, a, b, c, d))
             self.excitation_operator_type.append("quadruple")
             self.n_params += 1
 
@@ -530,7 +510,7 @@ class UccStructure:
             active_unocc_spin_idx: Active weakly occupied spin orbital indices.
         """
         for a, i, b, j, c, k, d, l, e, m in iterate_t5(active_occ_spin_idx, active_unocc_spin_idx):
-            self.excitation_indicies.append((i, j, k, l, m, a, b, c, d, e))
+            self.excitation_indices.append((i, j, k, l, m, a, b, c, d, e))
             self.excitation_operator_type.append("quintuple")
             self.n_params += 1
 
@@ -542,22 +522,26 @@ class UccStructure:
             active_unocc_spin_idx: Active weakly occupied spin orbital indices.
         """
         for a, i, b, j, c, k, d, l, e, m, f, n in iterate_t6(active_occ_spin_idx, active_unocc_spin_idx):
-            self.excitation_indicies.append((i, j, k, l, m, n, a, b, c, d, e, f))
+            self.excitation_indices.append((i, j, k, l, m, n, a, b, c, d, e, f))
             self.excitation_operator_type.append("sextuple")
             self.n_params += 1
 
 
 class UpsStructure:
+    __slots__ = ("excitation_indices", "excitation_operator_type", "n_params", "grad_param_R", "param_names")
+
     def __init__(self) -> None:
-        """Intialize the unitary product state ansatz structure."""
-        self.excitation_indicies: list[tuple[int, ...]] = []
+        """Initialize the unitary product state ansatz structure."""
+        self.excitation_indices: list[tuple[int, ...]] = []
         self.excitation_operator_type: list[str] = []
-        self.n_params = 0
+        self.n_params: int = 0
+        self.grad_param_R: dict[str, int] = {}
+        self.param_names: list[str] = []
 
     def create_tups(self, num_active_orbs: int, ansatz_options: dict[str, Any]) -> None:
         """Create tUPS ansatz.
 
-        #. 10.1103/PhysRevResearch.6.023300
+        #. 10.1103/PhysRevResearch.6.023300 (tUPS)
         #. 10.1088/1367-2630/ac2cb3 (QNP)
 
         Ansatz Options:
@@ -572,6 +556,7 @@ class UpsStructure:
         Returns:
             tUPS ansatz.
         """
+        # Options
         valid_options = ("n_layers", "do_qnp", "skip_last_singles")
         for option in ansatz_options:
             if option not in valid_options:
@@ -587,16 +572,21 @@ class UpsStructure:
             skip_last_singles = ansatz_options["skip_last_singles"]
         else:
             skip_last_singles = False
+        # Layer loop
         for n in range(n_layers):
-            for p in range(0, num_active_orbs - 1, 2):
+            for p in range(0, num_active_orbs - 1, 2):  # first column of brick-wall
                 if not do_qnp:
                     # First single
-                    self.excitation_operator_type.append("tups_single")
-                    self.excitation_indicies.append((p,))
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((p, p + 1))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
                     self.n_params += 1
                 # Double
-                self.excitation_operator_type.append("tups_double")
-                self.excitation_indicies.append((p,))
+                self.excitation_operator_type.append("double")
+                self.excitation_indices.append((2 * p, 2 * p + 1, 2 * p + 2, 2 * p + 3))
+                self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                self.param_names.append(f"p{self.n_params:09d}")
                 self.n_params += 1
                 # Second single
                 if n + 1 == n_layers and skip_last_singles and num_active_orbs == 2:
@@ -604,31 +594,47 @@ class UpsStructure:
                     # Here the layer is only one block, thus,
                     # the last single excitation is earlier than expected.
                     continue
-                self.excitation_operator_type.append("tups_single")
-                self.excitation_indicies.append((p,))
+                self.excitation_operator_type.append("sa_single")
+                self.excitation_indices.append((p, p + 1))
+                self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                self.param_names.append(f"p{self.n_params:09d}")
                 self.n_params += 1
-            for p in range(1, num_active_orbs - 1, 2):
+            for p in range(1, num_active_orbs - 1, 2):  # second column of brick-wall
                 if not do_qnp:
                     # First single
-                    self.excitation_operator_type.append("tups_single")
-                    self.excitation_indicies.append((p,))
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((p, p + 1))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
                     self.n_params += 1
                 # Double
-                self.excitation_operator_type.append("tups_double")
-                self.excitation_indicies.append((p,))
+                self.excitation_operator_type.append("double")
+                self.excitation_indices.append((2 * p, 2 * p + 1, 2 * p + 2, 2 * p + 3))
+                self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                self.param_names.append(f"p{self.n_params:09d}")
                 self.n_params += 1
                 # Second single
                 if n + 1 == n_layers and skip_last_singles:
                     continue
-                self.excitation_operator_type.append("tups_single")
-                self.excitation_indicies.append((p,))
+                self.excitation_operator_type.append("sa_single")
+                self.excitation_indices.append((p, p + 1))
+                self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                self.param_names.append(f"p{self.n_params:09d}")
                 self.n_params += 1
 
     def create_fUCC(self, num_orbs: int, num_elec: int, ansatz_options: dict[str, Any]) -> None:
-        """Create factorized UCCSD ansatz.
+        """Create factorized UCC ansatz.
+
+        #. 10.1021/acs.jctc.8b01004 (k-UpCCGSD)
 
         Ansatz Options:
             * n_layers [int]: Number of layers.
+            * S [bool]: Add single excitations.
+            * SAS [bool]: Add spin-adapted single excitations.
+            * SAGS [bool]: Add generalized spin-adapted single excitations.
+            * D [bool]: Add double excitations.
+            * pD [bool]: Add pair double excitations.
+            * GpD [bool]: Add generalized pair double excitations.
 
         Args:
             num_orbs: Number of active spatial orbitals.
@@ -636,14 +642,41 @@ class UpsStructure:
             ansatz_options: Ansatz options.
 
         Returns:
-            Factorized UCCSD ansatz.
+            Factorized UCC ansatz.
         """
-        valid_options = ("n_layers",)
+        # Options
+        valid_options = ("n_layers", "S", "D", "SAGS", "pD", "GpD", "SAS")
         for option in ansatz_options:
             if option not in valid_options:
                 raise ValueError(f"Got unknown option for fUCC, {option}. Valid options are: {valid_options}")
         if "n_layers" not in ansatz_options.keys():
             raise ValueError("fUCC require the option 'n_layers'")
+        do_S = False
+        do_SAS = False
+        do_SAGS = False
+        do_D = False
+        do_pD = False
+        do_GpD = False
+        if "S" in ansatz_options.keys():
+            if ansatz_options["S"]:
+                do_S = True
+        if "SAS" in ansatz_options.keys():
+            if ansatz_options["SAS"]:
+                do_SAS = True
+        if "SAGS" in ansatz_options.keys():
+            if ansatz_options["SAGS"]:
+                do_SAGS = True
+        if "D" in ansatz_options.keys():
+            if ansatz_options["D"]:
+                do_D = True
+        if "pD" in ansatz_options.keys():
+            if ansatz_options["pD"]:
+                do_pD = True
+        if "GpD" in ansatz_options.keys():
+            if ansatz_options["GpD"]:
+                do_GpD = True
+        if True not in (do_S, do_SAS, do_SAGS, do_D, do_pD, do_GpD):
+            raise ValueError("fUCC requires some excitations got none.")
         n_layers = ansatz_options["n_layers"]
         num_spin_orbs = 2 * num_orbs
         occ = []
@@ -655,54 +688,53 @@ class UpsStructure:
         for _ in range(num_spin_orbs - np.sum(num_elec)):
             unocc.append(idx)
             idx += 1
+        # Layer loop
         for _ in range(n_layers):
-            for a, i in iterate_t1(occ, unocc):
-                self.excitation_operator_type.append("single")
-                self.excitation_indicies.append((i, a))
-                self.n_params += 1
-            for a, i, b, j in iterate_t2(occ, unocc):
-                self.excitation_operator_type.append("double")
-                self.excitation_indicies.append((i, j, a, b))
-                self.n_params += 1
+            if do_S:
+                for a, i in iterate_t1(occ, unocc):
+                    self.excitation_operator_type.append("single")
+                    self.excitation_indices.append((i, a))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_SAS:
+                for a, i, _ in iterate_t1_sa(occ, unocc):
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i, a))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_SAGS:
+                for a, i, _ in iterate_t1_sa_generalized(num_orbs):
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i, a))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_D:
+                for a, i, b, j in iterate_t2(occ, unocc):
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_pD:
+                for a, i, b, j in iterate_pair_t2(occ, unocc):
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_GpD:
+                for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
 
-    def create_kSAfUpCCGSD(self, num_orbs: int, ansatz_options: dict[str, Any]) -> None:
-        """Create modified k-UpCCGSD ansatz.
-
-        The ansatz have been modifed to use spin-adapted singet single excitation operators.
-
-        #. 10.1021/acs.jctc.8b01004
-
-        Ansatz Options:
-            * n_layers [int]: Number of layers.
-
-        Args:
-            num_orbs: Number of spatial active orbitals.
-            ansatz_options: Ansatz options.
-
-        Returns:
-            Modified k-UpCCGSD ansatz.
-        """
-        valid_options = ("n_layers",)
-        for option in ansatz_options:
-            if option not in valid_options:
-                raise ValueError(
-                    f"Got unknown option for kSAfUpCCGSD, {option}. Valid options are: {valid_options}"
-                )
-        if "n_layers" not in ansatz_options.keys():
-            raise ValueError("kSAfUpCCGSD require the option 'n_layers'")
-        n_layers = ansatz_options["n_layers"]
-        for _ in range(n_layers):
-            for a, i, _ in iterate_t1_sa_generalized(num_orbs):
-                self.excitation_operator_type.append("sa_single")
-                self.excitation_indicies.append((i, a))
-                self.n_params += 1
-            for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
-                self.excitation_operator_type.append("double")
-                self.excitation_indicies.append((i, j, a, b))
-                self.n_params += 1
-
-    def create_dUCCSD(self, num_orbs: int, num_elec: int, ansatz_options: dict[str, Any]) -> None:
-        r"""Create disentangled UCCSD ansatz with SDS operator ordering.
+    def create_SDSfUCC(self, num_orbs: int, num_elec: int, ansatz_options: dict[str, Any]) -> None:
+        r"""Create SDS ordered factorized UCC.
 
         The operator ordering of this implementation is,
 
@@ -711,10 +743,14 @@ class UpsStructure:
             \exp\left(\theta_{ijab}\left(\hat{T}_{ijab}-\hat{T}_{ijab}^\dagger\right)\right)
             \exp\left(\theta_{ia}\left(\hat{T}_{ia}-\hat{T}_{ia}^\dagger\right)\right)\left|\text{CSF}\right>
 
-        #. 10.1063/1.5133059, Eq. 25, Eq. 35
+        #. 10.1063/1.5133059, Eq. 25, Eq. 35 (SDS)
+        #. 10.1021/acs.jctc.8b01004 (k-UpCCGSD)
 
         Ansatz Options:
             * n_layers [int]: Number of layers.
+            * D [bool]: Add double excitations.
+            * pD [bool]: Add pair double excitations.
+            * GpD [bool]: Add generalized pair double excitations.
 
         Args:
             num_orbs: Number of active spatial orbitals.
@@ -722,16 +758,31 @@ class UpsStructure:
             ansatz_options: Ansatz options.
 
         Returns:
-            Disentaglned UCCSD ansatz.
+            SDS ordered fUCC ansatz.
         """
-        valid_options = ("n_layers",)
+        # Options
+        valid_options = ("n_layers", "D", "pD", "GpD")
         for option in ansatz_options:
             if option not in valid_options:
                 raise ValueError(
-                    f"Got unknown option for dUCCSD, {option}. Valid options are: {valid_options}"
+                    f"Got unknown option for SDSfUCC, {option}. Valid options are: {valid_options}"
                 )
         if "n_layers" not in ansatz_options.keys():
-            raise ValueError("dUCCSD require the option 'n_layers'")
+            raise ValueError("SDSfUCC require the option 'n_layers'")
+        do_D = False
+        do_pD = False
+        do_GpD = False
+        if "D" in ansatz_options.keys():
+            if ansatz_options["D"]:
+                do_D = True
+        if "pD" in ansatz_options.keys():
+            if ansatz_options["pD"]:
+                do_pD = True
+        if "GpD" in ansatz_options.keys():
+            if ansatz_options["GpD"]:
+                do_GpD = True
+        if True not in (do_D, do_pD, do_GpD):
+            raise ValueError("SDSfUCC requires some excitations got none.")
         n_layers = ansatz_options["n_layers"]
         num_spin_orbs = 2 * num_orbs
         occ = []
@@ -743,54 +794,63 @@ class UpsStructure:
         for _ in range(num_spin_orbs - np.sum(num_elec)):
             unocc.append(idx)
             idx += 1
+        # Layer loop
         for _ in range(n_layers):
-            for a, i, b, j in iterate_t2(occ, unocc):
-                self.excitation_operator_type.append("single")
-                self.excitation_indicies.append((i, a))
-                self.n_params += 1
-                self.excitation_operator_type.append("double")
-                self.excitation_indicies.append((i, j, a, b))
-                self.n_params += 1
-                self.excitation_operator_type.append("single")
-                self.excitation_indicies.append((j, b))
-                self.n_params += 1
-
-    def create_kSAdUpCCGSD(self, num_orbs: int, ansatz_options: dict[str, Any]) -> None:
-        """Create modified k-UpCCGSD ansatz.
-
-        The ansatz have been modifed to use spin-adapted singet single excitation operators.
-        This ansatz has also been modified to follow the operator ordering as dUCCSD.
-
-        #. 10.1021/acs.jctc.8b01004
-        #. 10.1063/1.5133059, Eq. 25, Eq. 35 (for ordering)
-
-        Ansatz Options:
-            * n_layers [int]: Number of layers.
-
-        Args:
-            num_orbs: Number of spatial active orbitals.
-            ansatz_options: Ansatz options.
-
-        Returns:
-            Modified k-UpCCGSD ansatz.
-        """
-        valid_options = "n_layers"
-        for option in ansatz_options:
-            if option not in valid_options:
-                raise ValueError(
-                    f"Got unknown option for kSAdUpCCGSD, {option}. Valid options are: {valid_options}"
-                )
-        if "n_layers" not in ansatz_options.keys():
-            raise ValueError("kSAdUpCCGSD require the option 'n_layers'")
-        n_layers = ansatz_options["n_layers"]
-        for _ in range(n_layers):
-            for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
-                self.excitation_operator_type.append("sa_single")
-                self.excitation_indicies.append((i // 2, a // 2))
-                self.n_params += 1
-                self.excitation_operator_type.append("double")
-                self.excitation_indicies.append((i, j, a, b))
-                self.n_params += 1
-                self.excitation_operator_type.append("sa_single")
-                self.excitation_indicies.append((i // 2, a // 2))
-                self.n_params += 1
+            # Kind of D excitation determines indices for complete SDS block
+            if do_D:
+                for a, i, b, j in iterate_t2(occ, unocc):
+                    if i % 2 == a % 2:
+                        self.excitation_indices.append((i, a))
+                    else:
+                        self.excitation_indices.append((i, b))
+                    self.excitation_operator_type.append("single")
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    if i % 2 == a % 2:
+                        self.excitation_indices.append((j, b))
+                    else:
+                        self.excitation_indices.append((j, a))
+                    self.excitation_operator_type.append("single")
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_pD:
+                for a, i, b, j in iterate_pair_t2(occ, unocc):
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i // 2, a // 2))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i // 2, a // 2))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+            if do_GpD:
+                for a, i, b, j in iterate_pair_t2_generalized(num_orbs):
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i // 2, a // 2))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    self.excitation_operator_type.append("double")
+                    self.excitation_indices.append((i, j, a, b))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 2
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
+                    self.excitation_operator_type.append("sa_single")
+                    self.excitation_indices.append((i // 2, a // 2))
+                    self.grad_param_R[f"p{self.n_params:09d}"] = 4
+                    self.param_names.append(f"p{self.n_params:09d}")
+                    self.n_params += 1
