@@ -27,7 +27,6 @@ def test_H2_631g_naive():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -35,7 +34,7 @@ def test_H2_631g_naive():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = naive.LinearResponseUCC(WF, excitations="SD")
@@ -91,9 +90,8 @@ def test_LiH_sto3g_naive():
     SQobj.hartree_fock.run_restricted_hartree_fock()
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
-    # OO-UCCSD
+    # oo-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -101,7 +99,7 @@ def test_LiH_sto3g_naive():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = naive.LinearResponseUCC(WF, excitations="SD")
@@ -173,7 +171,6 @@ def test_H2_631g_projLR():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -181,7 +178,7 @@ def test_H2_631g_projLR():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = projected.LinearResponseUCC(WF, excitations="SD")
@@ -232,9 +229,8 @@ def test_LiH_sto3g_proj():
     SQobj.hartree_fock.run_restricted_hartree_fock()
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
-    # OO-UCCSD
+    # oo-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -242,7 +238,7 @@ def test_LiH_sto3g_proj():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     LR = projected.LinearResponseUCC(WF, excitations="SD")
     LR.calc_excitation_energies()
@@ -281,7 +277,7 @@ def test_LiH_sto3g_proj():
     assert abs(osc_strengths[5] - 0.010379091370812886) < thresh
     assert abs(osc_strengths[6] - 0.010379091373763447) < thresh
     assert abs(osc_strengths[7] - 0.006256710161922168) < thresh
-    assert abs(osc_strengths[8] - 0.0623868647774392) < thresh
+    assert abs(osc_strengths[8] - 0.062488043049451776) < thresh
     assert abs(osc_strengths[9] - 0.12886225822034553) < thresh
     assert abs(osc_strengths[10] - 0.12886225822019629) < thresh
     assert abs(osc_strengths[11] - 0.046007031170702296) < thresh
@@ -307,7 +303,6 @@ def test_H2_631g_STLR():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -315,7 +310,7 @@ def test_H2_631g_STLR():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = statetransfer.LinearResponseUCC(
@@ -370,7 +365,6 @@ def test_LiH_sto3g_st():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -378,7 +372,7 @@ def test_LiH_sto3g_st():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = statetransfer.LinearResponseUCC(
@@ -421,7 +415,7 @@ def test_LiH_sto3g_st():
     assert abs(osc_strengths[5] - 0.010379091370809963) < thresh
     assert abs(osc_strengths[6] - 0.010379091373763017) < thresh
     assert abs(osc_strengths[7] - 0.0062567030068973305) < thresh
-    assert abs(osc_strengths[8] - 0.06238684901394572) < thresh
+    assert abs(osc_strengths[8] - 0.06248802645793188) < thresh
     assert abs(osc_strengths[9] - 0.12886225822029365) < thresh
     assert abs(osc_strengths[10] - 0.12886225822018843) < thresh
     assert abs(osc_strengths[11] - 0.04600702378157588) < thresh
@@ -447,7 +441,6 @@ def test_H2_631g_allST():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -455,7 +448,7 @@ def test_H2_631g_allST():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = allstatetransfer.LinearResponseUCC(
@@ -506,7 +499,6 @@ def test_LiH_sto3g_allST():
     g_eri = SQobj.integral.electron_repulsion_tensor
     # OO-UCCSD
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -514,7 +506,7 @@ def test_LiH_sto3g_allST():
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     # Linear Response
     LR = allstatetransfer.LinearResponseUCC(
