@@ -2,11 +2,11 @@
 import numpy as np
 import pyscf
 from qiskit_aer.primitives import Sampler
-from qiskit_nature.second_q.mappers import JordanWignerMapper
+from qiskit_nature.second_q.mappers import JordanWignerMapper, ParityMapper
 
 import slowquant.SlowQuant as sq
+from slowquant.qiskit_interface.circuit_wavefunction import WaveFunctionCircuit
 from slowquant.qiskit_interface.interface import QuantumInterface
-from slowquant.qiskit_interface.wavefunction import WaveFunction as WaveFunctionQC
 from slowquant.unitary_coupled_cluster.ups_wavefunction import WaveFunctionUPS
 
 
@@ -27,7 +27,6 @@ def test_tups() -> None:
 
     # Conventional UPS wave function
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -37,7 +36,7 @@ def test_tups() -> None:
         ansatz_options={"n_layers": 1, "skip_last_singles": True},
         include_active_kappa=True,
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     assert abs(WF.energy_elec - -8.82891657651419) < 10**-8
 
@@ -47,16 +46,15 @@ def test_tups() -> None:
     QI = QuantumInterface(
         primitive, "tUPS", mapper, ansatz_options={"n_layers": 1, "skip_last_singles": True}
     )
-    qWF = WaveFunctionQC(
-        SQobj.molecule.number_bf * 2,
+    qWF = WaveFunctionCircuit(
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         QI,
     )
-    qWF.ansatz_parameters = WF.thetas
+    qWF.thetas = WF.thetas
 
     assert abs(qWF.energy_elec - -8.82891657651419) < 10**-8
 
@@ -78,17 +76,16 @@ def test_fucc() -> None:
 
     # Conventional UPS wave function
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
         h_core,
         g_eri,
-        "fUCC",
+        "fUCCSD",
         ansatz_options={},
         include_active_kappa=True,
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     assert abs(WF.energy_elec - -8.828916576513892) < 10**-8
 
@@ -96,16 +93,15 @@ def test_fucc() -> None:
     mapper = JordanWignerMapper()
     primitive = Sampler(run_options={"shots": None})
     QI = QuantumInterface(primitive, "fUCCSD", mapper, ansatz_options={})
-    qWF = WaveFunctionQC(
-        SQobj.molecule.number_bf * 2,
+    qWF = WaveFunctionCircuit(
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         QI,
     )
-    qWF.ansatz_parameters = WF.thetas
+    qWF.thetas = WF.thetas
 
     assert abs(qWF.energy_elec - -8.828916576513892) < 10**-8
 
@@ -127,7 +123,6 @@ def test_ksafupccgsd() -> None:
 
     # Conventional UPS wave function
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -137,7 +132,7 @@ def test_ksafupccgsd() -> None:
         ansatz_options={"n_layers": 1},
         include_active_kappa=True,
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     assert abs(WF.energy_elec - -8.828916576543133) < 10**-8
 
@@ -145,22 +140,21 @@ def test_ksafupccgsd() -> None:
     mapper = JordanWignerMapper()
     primitive = Sampler(run_options={"shots": None})
     QI = QuantumInterface(primitive, "kSAfUpCCGSD", mapper, ansatz_options={"n_layers": 1})
-    qWF = WaveFunctionQC(
-        SQobj.molecule.number_bf * 2,
+    qWF = WaveFunctionCircuit(
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         QI,
     )
-    qWF.ansatz_parameters = WF.thetas
+    qWF.thetas = WF.thetas
 
     assert abs(qWF.energy_elec - -8.828916576543133) < 10**-8
 
 
-def test_duccsd() -> None:
-    """Test dUCCSD is conventional <--> qiskit compatibile."""
+def test_sdsfuccsd() -> None:
+    """Test SDSfUCCSD is conventional <--> qiskit compatibile."""
     # Setup initial stuff
     SQobj = sq.SlowQuant()
     SQobj.set_molecule(
@@ -176,40 +170,38 @@ def test_duccsd() -> None:
 
     # Conventional UPS wave function
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
         h_core,
         g_eri,
-        "dUCCSD",
+        "SDSfUCCSD",
         ansatz_options={},
         include_active_kappa=True,
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     assert abs(WF.energy_elec - -8.82891657653415) < 10**-8
 
     # Circuit based UPS wave function
     mapper = JordanWignerMapper()
     primitive = Sampler(run_options={"shots": None})
-    QI = QuantumInterface(primitive, "dUCCSD", mapper, ansatz_options={})
-    qWF = WaveFunctionQC(
-        SQobj.molecule.number_bf * 2,
+    QI = QuantumInterface(primitive, "SDSfUCCSD", mapper, ansatz_options={})
+    qWF = WaveFunctionCircuit(
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         QI,
     )
-    qWF.ansatz_parameters = WF.thetas
+    qWF.thetas = WF.thetas
 
     assert abs(qWF.energy_elec - -8.82891657653415) < 10**-8
 
 
-def test_ksadupccgsd() -> None:
-    """Test k-SAdUpCCGSD is conventional <--> qiskit compatibile."""
+def test_ksasdsfupccgsd() -> None:
+    """Test k-SASDSfUpCCGSD is conventional <--> qiskit compatibile."""
     # Setup initial stuff
     SQobj = sq.SlowQuant()
     SQobj.set_molecule(
@@ -225,34 +217,32 @@ def test_ksadupccgsd() -> None:
 
     # Conventional UPS wave function
     WF = WaveFunctionUPS(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
         h_core,
         g_eri,
-        "kSAdUpCCGSD",
+        "kSASDSfUpCCGSD",
         ansatz_options={"n_layers": 1},
         include_active_kappa=True,
     )
-    WF.run_ups(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     assert abs(WF.energy_elec - -8.828916576542285) < 10**-8
 
     # Circuit based UPS wave function
     mapper = JordanWignerMapper()
     primitive = Sampler(run_options={"shots": None})
-    QI = QuantumInterface(primitive, "kSAdUpCCGSD", mapper, ansatz_options={"n_layers": 1})
-    qWF = WaveFunctionQC(
-        SQobj.molecule.number_bf * 2,
+    QI = QuantumInterface(primitive, "kSASDSfUpCCGSD", mapper, ansatz_options={"n_layers": 1})
+    qWF = WaveFunctionCircuit(
         SQobj.molecule.number_electrons,
         (2, 2),
-        WF.c_trans,
+        WF.c_mo,
         h_core,
         g_eri,
         QI,
     )
-    qWF.ansatz_parameters = WF.thetas
+    qWF.thetas = WF.thetas
 
     assert abs(qWF.energy_elec - -8.828916576542285) < 10**-8
 
@@ -270,24 +260,22 @@ def test_lih_fucc_allparameters() -> None:
 
     # SlowQuant
     WF = WaveFunctionUPS(
-        mol.nao * 2,
         mol.nelectron,
         (2, 3),
         rhf.mo_coeff,
         mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
         mol.intor("int2e"),
-        "fUCC",
+        "fUCCSD",
     )
     sampler = Sampler()
     mapper = JordanWignerMapper()
 
     QI = QuantumInterface(sampler, "fUCCSD", mapper)
 
-    qWF = WaveFunctionQC(
-        mol.nao * 2,
+    qWF = WaveFunctionCircuit(
         mol.nelectron,
         (2, 3),
-        WF.c_trans,
+        WF.c_mo,
         mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
         mol.intor("int2e"),
         QI,
@@ -296,5 +284,153 @@ def test_lih_fucc_allparameters() -> None:
         thetas = np.zeros(len(WF.thetas))
         thetas[n] = 1
         WF.thetas = thetas
-        qWF.ansatz_parameters = thetas
-        assert abs(qWF.energy_elec - WF.energy_elec) < 10**-12
+        qWF.thetas = thetas
+        assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+
+def test_lih_fucc_mappings() -> None:
+    """Tests that parity mapping and jw mapping gives the same for fUCC type wave function."""
+    atom = "Li .0 .0 .0; H .0 .0 1.672"
+    basis = "sto3g"
+
+    # PySCF
+    mol = pyscf.M(atom=atom, basis=basis, unit="angstrom")
+    rhf = pyscf.scf.RHF(mol).run()
+
+    # SlowQuant
+    WF = WaveFunctionUPS(
+        mol.nelectron,
+        (2, 5),
+        rhf.mo_coeff,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        "fUCCSD",
+    )
+    assert abs(WF.energy_elec - -8.808220920154708) < 10**-10
+
+    sampler = Sampler()
+    mapper = JordanWignerMapper()
+    QI = QuantumInterface(sampler, "fUCCSD", mapper)
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+    sampler = Sampler()
+    mapper = ParityMapper((1, 1))
+    QI = QuantumInterface(sampler, "fUCCSD", mapper)
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+
+def test_lih_sdsfucc_mappings() -> None:
+    """Tests that parity mapping and jw mapping gives the same for SDSfUCC type wave function."""
+    atom = "Li .0 .0 .0; H .0 .0 1.672"
+    basis = "sto3g"
+
+    # PySCF
+    mol = pyscf.M(atom=atom, basis=basis, unit="angstrom")
+    rhf = pyscf.scf.RHF(mol).run()
+
+    # SlowQuant
+    WF = WaveFunctionUPS(
+        mol.nelectron,
+        (2, 5),
+        rhf.mo_coeff,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        "SDSfUCCSD",
+    )
+    assert abs(WF.energy_elec - -8.808220920154705) < 10**-10
+
+    sampler = Sampler()
+    mapper = JordanWignerMapper()
+    QI = QuantumInterface(sampler, "SDSfUCCSD", mapper)
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+    sampler = Sampler()
+    mapper = ParityMapper((1, 1))
+    QI = QuantumInterface(sampler, "SDSfUCCSD", mapper)
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+
+def test_lih_tups_mappings() -> None:
+    """Tests that parity mapping and jw mapping gives the same for tUPS type wave function."""
+    atom = "Li .0 .0 .0; H .0 .0 1.672"
+    basis = "sto3g"
+
+    # PySCF
+    mol = pyscf.M(atom=atom, basis=basis, unit="angstrom")
+    rhf = pyscf.scf.RHF(mol).run()
+
+    # SlowQuant
+    WF = WaveFunctionUPS(
+        mol.nelectron,
+        (2, 5),
+        rhf.mo_coeff,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        "tUPS",
+        ansatz_options={"n_layers": 4},
+    )
+    assert abs(WF.energy_elec - -8.808220920154707) < 10**-10
+
+    sampler = Sampler()
+    mapper = JordanWignerMapper()
+    QI = QuantumInterface(sampler, "tUPS", mapper, ansatz_options={"n_layers": 4})
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
+
+    sampler = Sampler()
+    mapper = ParityMapper((1, 1))
+    QI = QuantumInterface(sampler, "tUPS", mapper, ansatz_options={"n_layers": 4})
+    qWF = WaveFunctionCircuit(
+        mol.nelectron,
+        (2, 5),
+        WF.c_mo,
+        mol.intor("int1e_kin") + mol.intor("int1e_nuc"),
+        mol.intor("int2e"),
+        QI,
+    )
+    qWF.thetas = WF.thetas
+    assert abs(qWF.energy_elec - WF.energy_elec) < 10**-10
