@@ -27,7 +27,6 @@ def test_h2_sto3g_uccsd_lr() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -35,7 +34,7 @@ def test_h2_sto3g_uccsd_lr() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(False)
+    WF.run_wf_optimization_1step("SLSQP", False)
     # SC
     LR = allselfconsistent.LinearResponseUCC(
         WF,
@@ -69,7 +68,6 @@ def test_LiH_atmethods_energies() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -77,7 +75,7 @@ def test_LiH_atmethods_energies() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     threshold = 10 ** (-3)
 
@@ -143,7 +141,6 @@ def test_LiH_naiveq_methods_energies() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -151,7 +148,7 @@ def test_LiH_naiveq_methods_energies() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     threshold = 10 ** (-5)
 
@@ -256,62 +253,6 @@ def test_LiH_naiveq_methods_energies() -> None:
     )
     assert np.allclose(LR_naive.excitation_energies, solutions, atol=threshold)
 
-    # HST
-    LR_HST_naive = statetransfer.LinearResponseUCC(
-        WF,
-        excitations="SD",
-        do_approximate_hermitification=True,
-    )
-    LR_HST_naive.calc_excitation_energies()
-
-    solutions = np.array(
-        [
-            0.12957234,
-            0.17886086,
-            0.17886086,
-            0.60515489,
-            0.64717441,
-            0.74104045,
-            0.74104045,
-            1.003942,
-            2.07479277,
-            2.13715595,
-            2.13715595,
-            2.45576414,
-            2.95517029,
-        ]
-    )
-
-    assert np.allclose(LR_HST_naive.excitation_energies, solutions, atol=threshold * 10)
-
-    # HSC
-    LR_HSC_naive = selfconsistent.LinearResponseUCC(
-        WF,
-        excitations="SD",
-        do_approximate_hermitification=True,
-    )
-    LR_HSC_naive.calc_excitation_energies()
-
-    solutions = np.array(
-        [
-            0.12957234,
-            0.17886086,
-            0.17886086,
-            0.60515489,
-            0.64717441,
-            0.74104045,
-            0.74104045,
-            1.003942,
-            2.07479277,
-            2.13715595,
-            2.13715595,
-            2.45576414,
-            2.95517029,
-        ]
-    )
-
-    assert np.allclose(LR_HSC_naive.excitation_energies, solutions, atol=threshold * 10)
-
 
 def test_LiH_naiveq_methods_matrices() -> None:
     """Test LiH all matrices and their properties for naive q LR methods."""
@@ -328,7 +269,6 @@ def test_LiH_naiveq_methods_matrices() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -336,7 +276,7 @@ def test_LiH_naiveq_methods_matrices() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     threshold = 10 ** (-5)
 
@@ -382,16 +322,6 @@ def test_LiH_naiveq_methods_matrices() -> None:
     assert np.allclose(LR_naive.B, LR_naive.B.T, atol=threshold)
     assert np.allclose(LR_naive.Delta, np.zeros_like(LR_naive.Delta), atol=threshold)
 
-    # HST
-    LR_HST_naive = statetransfer.LinearResponseUCC(
-        WF,
-        excitations="SD",
-        do_approximate_hermitification=True,
-    )
-
-    assert np.allclose(LR_HST_naive.A, LR_HST_naive.A.T, atol=threshold)
-    assert np.allclose(LR_HST_naive.B, LR_HST_naive.B.T, atol=threshold)
-
 
 def test_LiH_allproj_energies() -> None:
     """Test LiH for all-proj LR method."""
@@ -408,7 +338,6 @@ def test_LiH_allproj_energies() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -416,7 +345,7 @@ def test_LiH_allproj_energies() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     threshold = 10 ** (-5)
 
@@ -461,7 +390,6 @@ def test_LiH_STproj_energies() -> None:
     h_core = SQobj.integral.kinetic_energy_matrix + SQobj.integral.nuclear_attraction_matrix
     g_eri = SQobj.integral.electron_repulsion_tensor
     WF = WaveFunctionUCC(
-        SQobj.molecule.number_bf * 2,
         SQobj.molecule.number_electrons,
         (2, 2),
         SQobj.hartree_fock.mo_coeff,
@@ -469,7 +397,7 @@ def test_LiH_STproj_energies() -> None:
         g_eri,
         "SD",
     )
-    WF.run_ucc(True)
+    WF.run_wf_optimization_1step("SLSQP", True)
 
     threshold = 10 ** (-5)
 
@@ -498,31 +426,3 @@ def test_LiH_STproj_energies() -> None:
     )
 
     assert np.allclose(LR_naive.excitation_energies, solutions, atol=threshold)
-
-    # HSTproj
-    LR_HSTproj_naive = projected_statetransfer.LinearResponseUCC(
-        WF,
-        excitations="SD",
-        do_approximate_hermitification=True,
-    )
-    LR_HSTproj_naive.calc_excitation_energies()
-
-    solutions = np.array(
-        [
-            0.12972911,
-            0.18092787,
-            0.18092787,
-            0.6053739,
-            0.64747858,
-            0.74982806,
-            0.74982806,
-            1.00424528,
-            2.07489505,
-            2.13720683,
-            2.13720683,
-            2.45602345,
-            2.95608694,
-        ]
-    )
-
-    assert np.allclose(LR_HSTproj_naive.excitation_energies, solutions, atol=threshold * 10)  # type: ignore [assignment]
